% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aicc.R
\name{aicc}
\alias{aicc}
\title{AICc}
\usage{
aicc(model, env, parallel = FALSE)
}
\arguments{
\item{model}{Maxent object.}

\item{parallel}{logical, if TRUE it uses parallel computation, deafult is FALSE.}
}
\value{
The computed AICc
}
\description{
AICc
}
\examples{
\dontrun{
aicc(model, predictors, parallel = T)}

}
\author{
Sergio Vignali
}
