% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANN-class.R
\docType{class}
\name{ANN-class}
\alias{ANN-class}
\alias{ANN}
\title{Artificial Neural Network}
\description{
This Class represents an Artificial Neural Network model objects and hosts
all the information related to the model.
}
\details{
See \code{\link{nnet}} for the meaning of the slots.
}
\section{Slots}{

\describe{
\item{\code{size}}{integer. Number of the units in the hidden layer.}

\item{\code{decay}}{numeric. Weight decay.}

\item{\code{rang}}{numeric. Initial random weights.}

\item{\code{maxit}}{integer. Maximum number of iterations.}

\item{\code{model}}{\code{\link{nnet}}. The randomForest model object.}
}}

\author{
Sergio Vignali
}
