% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkMaxentInstallation.R
\name{checkMaxentInstallation}
\alias{checkMaxentInstallation}
\title{Check Maxent Installation}
\usage{
checkMaxentInstallation(verbose = TRUE)
}
\arguments{
\item{verbose}{logical, if \code{TRUE} the function provides useful messages to
understand what is not correctly installed, default is \code{TRUE}.}
}
\value{
\code{TRUE} if Maxent is correctly installed, \code{FALSE} otherwise.
}
\description{
The function checks if Maxent is correctly installed.
}
\details{
In order to have Maxent correctly configured is necessary that:
\itemize{
\item Java is installed;
\item the package "rJava" is installed;
\item the file "maxent.jar" is in the correct folder.
}
}
\examples{
checkMaxentInstallation()
}
\author{
Sergio Vignali
}
