% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{clamp}
\alias{clamp}
\title{Clamp}
\usage{
clamp(x, lower, upper)
}
\arguments{
\item{x}{numeric. Vector to be clamped.}

\item{lower}{numeric. The lowest value used to clamp.}

\item{upper}{numeric. The highest value used to clamp}
}
\value{
numeric. The clamped vector.
}
\description{
Clamp a given vector between the provided lower and upper values.
}
\details{
The code is inspired by
\url{http://gallery.rcpp.org/articles/sugar-function-clamp/}.
}
\examples{
clamp(1:20, 4, 17)

}
\author{
Sergio Vignali
}
