% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confMatrix.R
\name{confMatrix}
\alias{confMatrix}
\title{Confusion Matrix}
\usage{
confMatrix(model, test = NULL, th = NULL)
}
\arguments{
\item{model}{SDMmodel object.}

\item{test}{SWD test locations, if not provided it uses the train dataset,
default is NULL.}

\item{th}{vector if provided it computes the evaluation at the given thresholds,
default is NULL.}
}
\value{
The Confusion Matrix for all the used thresholds.
}
\description{
Computes Confusion Matrixes for threshold values varying from 0 to 1
}
\examples{
\dontrun{
confMatrix(my_model, test = test_dataset)}
}
