% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertFolds.R
\name{convertFolds}
\alias{convertFolds}
\title{Convert Folds}
\usage{
convertFolds(x, data)
}
\arguments{
\item{x}{list. Folds to be converted.}

\item{data}{\code{\linkS4class{SWD}} object used to create the folds.}
}
\value{
list with two matrices, the first for the training and the second for
the testing dataset. Each column of one matrix represents a fold with
\code{TRUE} for the locations included in and \code{FALSE} excluded from the
partition.
}
\description{
Convert folds created with other packages into a format valid for
\pkg{SDMtune}.
}
\details{
The function can convert folds created with the following packages:
ENMeval and CVblocks.
}
\examples{
\donttest{
# Acquire environmental variables
files <- list.files(path = file.path(system.file(package = "dismo"), "ex"),
                    pattern = "grd", full.names = TRUE)
predictors <- raster::stack(files)

# Prepare presence and background locations
p_coords <- virtualSp$presence
bg_coords <- virtualSp$background

data <- prepareSWD(species = "Virtual species", p = p_coords, a = bg_coords,
                   env = predictors, categorical = "biome")

# Example using the ENMeval package
# Block partition
# library(ENMeval)
block_folds <- get.block(occ = data@data[data@pa == 1, ],
                         bg.coords = data@data[data@pa == 0, ])
folds <- convertFolds(block_folds, data)
# Checkerboard1 partition
cb_folds <- get.checkerboard1(occ = data@data[data@pa == 1, ],
                              env = predictors,
                              bg.coords = data@data[data@pa == 0, ],
                              aggregation.factor = 4)
folds <- convertFolds(cb_folds, data)

# Exaple using blockCV package
# Environmental block
# library(blockCV)
# Create spatial points data frame
library(raster)
sp_df <- SpatialPointsDataFrame(data@coords, data = as.data.frame(data@pa),
                                proj4string = crs(predictors))
block_folds <- envBlock(rasterLayer = predictors,
                        speciesData = sp_df,
                        species = "data@pa",
                        k = 4,
                        standardization = "standard",
                        rasterBlock = FALSE,
                        numLimit = 100)
folds <- convertFolds(block_folds, data)
}
}
\references{
Muscarella, R., Galante, P.J., Soley-Guardia, M., Boria, R.A.,
Kass, J., Uriarte, M. and R.P. Anderson (2014). ENMeval: An R package for
conducting spatially independent evaluations and estimating optimal model
complexity for ecological niche models. Methods in Ecology and Evolution.

Roozbeh Valavi, Jane Elith, José Lahoz-Monfort and Gurutzeta Guillera-Arroita
(2018). blockCV: Spatial and environmental blocking for k-fold
cross-validation. \url{https://github.com/rvalavi/blockCV}.
}
\author{
Sergio Vignali
}
