% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corVar.R
\name{corVar}
\alias{corVar}
\title{Print Correlated Variables}
\usage{
corVar(bg, method = "spearman", cor_th = NULL, order = TRUE,
  remove_diagonal = TRUE)
}
\arguments{
\item{bg}{\link{SWD}. Locations used to test the correlation between
environmental variables.}

\item{method}{character. The method used to compute the correlation matrix,
default is "spearman".}

\item{cor_th}{numeric. If provided it prints only the variables whose
correlation coefficient is higher or lower than the given threshold, default
is NULL.}

\item{order}{logical, if TRUE the variable are ordered from the most to the
less highly correlated, default is TRUE.}

\item{remove_diagonal}{logical, if TRUE the values in the diagonal are,
removed, default is TRUE.}
}
\value{
The name of the correlated variables.
}
\description{
Utility that prints the name of correlated variables and the relative
correlation coefficient value.
}
\examples{
\dontrun{
corVar(bg, method = 'pearson', cor_th = 0.6)}

}
\author{
Sergio Vignali
}
