% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doJk.R
\name{doJk}
\alias{doJk}
\title{Jackknife Test}
\usage{
doJk(model, variables = NULL, with_only = TRUE, return_models = FALSE)
}
\arguments{
\item{model}{Maxent object..}

\item{variables}{vector. Variables used for the test, if not provided it takes all the variables
used to train the model, default is NULL.}

\item{with_only}{logical. If TRUE it runs the test also for each variable in isolation, default is TRUE.}

\item{return_models}{logical, If TRUE returns all the models together with the test result, default is FALSE.}
}
\value{
A data frame with the test results. If \strong{return_model = TRUE} it returns a list with
the data frame together with the models.
}
\description{
Run the Jackknife test for variable importance removing one variable at time and returns the train AUC.
If the model has a test dataset, it returns also the test AUC.
}
\examples{
\dontrun{
doJk(model, variable = c('bio1', 'bio12'), with_only = TRUE)}

}
\author{
Sergio Vignali
}
