% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulaFromLambdas.R
\name{formulaFromLambdas}
\alias{formulaFromLambdas}
\title{Formula From Lambdas}
\usage{
formulaFromLambdas(l)
}
\arguments{
\item{l}{The lambdas parameters.}
}
\value{
The formula to predict Maxent models.
}
\description{
Create the formula to predict Maxent models starting from the lambdas object.
Used internally to create the MaxentModel object.
}
\examples{
\dontrun{
formulaFromLambdas(lambda)}

}
\author{
Sergio Vignali
}
