% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NN_utils.R
\name{get_input_units}
\alias{get_input_units}
\title{Get Input Units}
\usage{
get_input_units(model)
}
\arguments{
\item{model}{NN object.}
}
\value{
numeric. The number of input units.
}
\description{
Get the number of input units in a NN model.
}
\examples{
\dontrun{get_input_units(model)}

}
\author{
Sergio Vignali
}
