% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_tunable_args}
\alias{get_tunable_args}
\title{Get Tunable Arguments}
\usage{
get_tunable_args(model)
}
\arguments{
\item{model}{\link{SDMmodel} or \link{SDMmodelCV} object.}
}
\value{
character vector
}
\description{
Returns the name of all function arguments that can be tuned for a given
model.
}
\examples{
\dontrun{get_tunable_args(my_model)}

}
\author{
Sergio Vignali
}
