% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFeatureArgs.R
\name{getFeatureArgs}
\alias{getFeatureArgs}
\title{Get Feature Arguments}
\usage{
getFeatureArgs(fc)
}
\arguments{
\item{fc}{The value of the feature combination, possible values are combinations of
"l", "q", "p", "h" and "t".}
}
\value{
A vector with the arguments.
}
\description{
Given a short word for the feature combination return the corresponding arguments used by MaxEnt.
}
\examples{
fc_args <- getFeatureArgs('lqp')

}
\author{
Sergio Vignali
}
