% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLambdas.R
\name{getLambdas}
\alias{getLambdas}
\title{Get Lambdas}
\usage{
getLambdas(lambda_file, bg)
}
\arguments{
\item{lambda_file}{The path of the \strong{.lambdas} file.}

\item{bg}{The background locations used to train the model, given as MaxentSWD object.}
}
\value{
The data fetched from the \strong{.lambdas} file.
}
\description{
Get lambdas object from \strong{.lambdas} file, Used internally to create the MaxentModel object.
}
\examples{
\dontrun{
getLambdas("species.lambdas"), bg)}

}
\author{
Sergio Vignali
}
