% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSubsample.R
\name{getSubsample}
\alias{getSubsample}
\title{Get Dataset Subsample}
\usage{
getSubsample(dataset, size, seed = NULL)
}
\arguments{
\item{dataset}{A MaxentSWD object.}

\item{size}{The size of the sub sample.}

\item{seed}{The value used to set the seed in order to have consistent results, default is NULL.}
}
\value{
The sub sample as MaxentSWD object.
}
\description{
Get a random subsample of a MaxentSWD object.
}
\examples{
\dontrun{
getSubsample(bg, value = 4500)}

}
\author{
Sergio Vignali
}
