% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridSearch.R
\name{gridSearch}
\alias{gridSearch}
\title{Grid Search}
\usage{
gridSearch(model, hypers, metric, test = NULL, bg4test = NULL,
  env = NULL, parallel = FALSE, seed = NULL)
}
\arguments{
\item{seed}{}
}
\description{
Grid Search
}
