% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NN_utils.R
\name{loadNN}
\alias{loadNN}
\title{Load NN}
\usage{
loadNN(file_name)
}
\arguments{
\item{file_name}{The name of the files without the extension in  which is
saved the object. The function expects to have the file in the working
directory.}
}
\value{
The created SDMmodel object containing the NN model.
}
\description{
Load a NN model in R from the files saved in the working directory.
}
\examples{
\dontrun{x <- loadNN("my_model")}

}
\author{
Sergio Vignali
}
