% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeArgs.R
\name{makeArgs}
\alias{makeArgs}
\title{Make Arguments}
\usage{
makeArgs(rm, fc, output_format = c("logistic", "cloglog", "raw"),
  test = NULL, response_curves = FALSE, iterations = 500,
  extra_args = NULL)
}
\arguments{
\item{rm}{The value of the regularization multiplier.}

\item{fc}{The value of the feature combination, possible values are combinations of
"l", "q", "p", "h" and "t".}

\item{output_format}{The MaxEnt output format, possible values are "logistic", "cloglog" and "raw"
or a shorten version of the previous valus, e.g. "log" for logistic.}

\item{test}{The path of the test file in a SWD format, if NULL is not used.}

\item{response_curves}{Flag to compute the response curves, default is FALSE.}

\item{iterations}{Number of iterations used by the Maxent alghoritm, default is 500.}

\item{extra_args}{Extra arguments used to run MAxEnt.}
}
\value{
A vactor with the arguments.
}
\description{
Build the arguments used to run a MaxEnt model.
}
\examples{
args <- makeArgs(rm = 2, fc = "LQ")

}
\author{
Sergio Vignali
}
