% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMaxent.R
\docType{class}
\name{Maxent-class}
\alias{Maxent-class}
\alias{Maxent}
\title{Maxent}
\description{
This Class represents a MaxEnt model objects and hosts all the information related to the model.
}
\section{Slots}{

\describe{
\item{\code{presence}}{SWD. The presence locations used to train the model.}

\item{\code{background}}{SWD. The backgorund locations used to train the model.}

\item{\code{test}}{SWD. The test locations used to validate the model.}

\item{\code{results}}{matrix. The result that usually MaxEnt provide as a csv file.}

\item{\code{rm}}{numeric. The value of the regularization multiplier used to train the model.}

\item{\code{fc}}{character. The feature class combination used to train the model.}

\item{\code{iterations}}{numeric. The number of iterations used to train the model.}

\item{\code{output_format}}{character. The output format of the model.}

\item{\code{lambdas}}{vector. The lambdas parameters of the model.}

\item{\code{coeff}}{data.frame. The lambda coefficients of the model.}

\item{\code{formula}}{formula. The formula used to make prediction.}

\item{\code{lpn}}{numeric. Linear Predictor Normalizer.}

\item{\code{dn}}{numeric. Density Normalizer.}

\item{\code{entropy}}{numeric. The entropy value.}

\item{\code{min_max}}{data.frame. The minimum and maximum values of the continuous variables, used for clamping.}

\item{\code{plot_data}}{list. If available a list containing the values used to plot the response curves.}

\item{\code{folder}}{character. The path for the folder where are saved all the files produced by MaxEnt,
available if the "folder" parameter is provided to the runMaxent function.}
}}

\author{
Sergio Vignali
}
