% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Maxent2MaxEnt.R
\name{Maxent2MaxEnt}
\alias{Maxent2MaxEnt}
\title{Maxent2MaxEnt}
\usage{
Maxent2MaxEnt(model)
}
\arguments{
\item{model}{The Maxent object to be converted.}
}
\value{
The converted MaxEnt object.
}
\description{
Converts a Maxent object to a dismo MaxEnt object.
}
\examples{
\dontrun{MaxentModel2MaxEnt(model)}

}
\author{
Sergio Vignali
}
