% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxentOutput.R
\name{maxentOutput}
\alias{maxentOutput}
\title{MaxEnt Output}
\usage{
maxentOutput(model, response_curves = FALSE, jk = FALSE, threads = 1,
  folder = NULL)
}
\arguments{
\item{model}{Maxent object.}

\item{response_curves}{logical, if TRUE it plots the response curves in the html output,
default is FALSE.}

\item{jk}{logical, if TRUE it runs the jackknife test, default FALSE.}

\item{threads}{numeric. Number of threads used by MaxEnt to compute the
Jackknife test, default is 1.}

\item{folder}{character. The name of the folder in which to save the MaxEnt output,
default is NULL meaning a temporary folder. The folder is created in the working directory.}
}
\description{
Given a model the function produces the output in the given folder.
}
\examples{
\dontrun{model <- produceOutput(model, response_curves = T, jacckinfe = T,
thread = 4, folder = 'my_folder'))}

}
\author{
Sergio Vignali
}
