% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Maxnet_class.R
\docType{class}
\name{Maxnet-class}
\alias{Maxnet-class}
\alias{Maxnet}
\title{Maxnet}
\description{
This Class represents a Maxnet model objects and hosts all the information related to the model.
}
\section{Slots}{

\describe{
\item{\code{presence}}{SWD. The presence locations used to train the model.}

\item{\code{background}}{SWD. The backgorund locations used to train the model.}

\item{\code{test}}{SWD. The test locations used to validate the model.}

\item{\code{rm}}{numeric. The value of the regularization multiplier used to train the model.}

\item{\code{fc}}{character. The feature class combination used to train the model.}

\item{\code{type}}{character. The output format of the model.}

\item{\code{lambdas}}{vector. The lambdas parameters of the model.}

\item{\code{coeff}}{data.frame. The lambda coefficients of the model.}

\item{\code{formula}}{formula. The formula used to make prediction.}

\item{\code{entropy}}{numeric. The entropy value.}

\item{\code{min_max}}{data.frame. The minimum and maximum values of the continuous variables, used for clamping.}
}}

\author{
Sergio Vignali
}
