% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeSWD.R
\name{mergeSWD}
\alias{mergeSWD}
\title{Merge SWD Objects}
\usage{
mergeSWD(swd1, swd2)
}
\arguments{
\item{swd1}{SWD object.}

\item{swd2}{SWD object.}
}
\value{
The merged SWD object.
}
\description{
Merge two SWD object.
}
\details{
In case the two SWD objects have different columns, only the common
columns are used in the merge object.
}
\examples{
\dontrun{
mergeSWD(train, background)}

}
\author{
Sergio Vignali
}
