% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelOutput.R
\name{modelOutput}
\alias{modelOutput}
\title{Model Output}
\usage{
modelOutput(model, type = c("cloglog", "logistic", "raw"), folder,
  test = NULL, response_curves = FALSE, jk = FALSE, threads = 1,
  extra_args = c("noaddsamplestobackground", "removeduplicates=false"))
}
\arguments{
\item{model}{SDMmodel object.}

\item{type}{character MaxEnt output type, possible values are "cloglog",
"logistic" and "raw", default is "cloglog".}

\item{folder}{character. The name of the folder in which to save the MaxEnt
output. The folder is created in the working directory.}

\item{test}{SWD object with the test locations, default is NULL.}

\item{response_curves}{logical, if TRUE it plots the response curves in the
html output, default is FALSE.}

\item{jk}{logical, if TRUE it runs the jackknife test, default FALSE.}

\item{threads}{numeric. Number of threads used by MaxEnt to compute the
Jackknife test, default is 1.}

\item{extra_args}{vector. Extra arguments used to run MaxEnt, e.g.
"removeduplicates=false", default is NULL.}

\item{extra_args}{vector. Extra arguments used to run MaxEnt, defalt is
c("noaddsamplestobackground", "removeduplicates=false").}
}
\value{
The SDMmodel object.
}
\description{
Given a model the function produces the output in the given folder.
}
\details{
The function by default \strong{uses extra_args =
c("noaddsamplestobackground", "removeduplicates=false")}. In case this is
not your expected beaviour you can remove both passing extra_args = "" or you
ca add any other additional arguments extending the previous vector.
}
\examples{
\dontrun{model <- modelOutput(model, type = "logistic", test = test,
response_curves = T, jk = T, thread = 4, folder = "my_folder"))}

}
\author{
Sergio Vignali
}
