% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelReport.R
\name{modelReport}
\alias{modelReport}
\title{Model Report}
\usage{
modelReport(model, type, folder, test = NULL, response_curves = FALSE,
  jk = FALSE, env = NULL, clamp = TRUE, permut = 10)
}
\arguments{
\item{model}{\link{SDMmodel} object.}

\item{type}{character. Output type, see \link{predict,Maxent-method} for
\link{Maxent} models or \link{predict.maxnet} for \link{Maxnet} models.}

\item{folder}{character. The name of the folder in which to save the output.
The folder is created in the working directory.}

\item{test}{\link{SWD} object with the test locations, default is NULL.}

\item{response_curves}{logical, if TRUE it plots the response curves in the
html output, default is FALSE.}

\item{jk}{logical, if TRUE it runs the jackknife test, default FALSE.}

\item{env}{\link{stack}. If provided it computes and adds a prediction map to
the output, default is NULL.}

\item{clamp}{logical for clumping during prediction, used for response curves
and for the prediction map, default is TRUE.}

\item{permut}{integer. Number of permutations, default is 10.}
}
\description{
Make a report that shows the main results.
}
\details{
The function produces a report similar to the one created by MaxEnt
software.
}
\examples{
\dontrun{
modelReport(model, "cloglog", folder = "my_folder", response_curves = T)}

}
\author{
Sergio Vignali
}
