% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NN_class.R
\docType{class}
\name{NN-class}
\alias{NN-class}
\alias{NN}
\title{NN}
\description{
This Class represents a Neural Network model objects and hosts all the information
related to the model.
}
\section{Slots}{

\describe{
\item{\code{model}}{keras.engine.sequential.Sequential. The Neural Network object.}

\item{\code{loss}}{character. Name of the loss function optimized.}

\item{\code{optimizer}}{character. Name of optimizer algorithm used to train the
model.}

\item{\code{min_max}}{data.frame. The minimum and maximum values of the continuous
variables, used for clamping.}

\item{\code{means}}{numeric. Vector of means used to standardize continuous
variables.}

\item{\code{stds}}{numeric. Vector of standard deviations used to standardize
continuous variables.}

\item{\code{levels}}{list with the levels of each categorical variable.}

\item{\code{yaml}}{character. The yaml representation of the model configuration.}
}}

\author{
Sergio Vignali
}
