% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help-functions.R
\name{old2NewSDMmodel}
\alias{old2NewSDMmodel}
\title{Convert Old SDMmodel object}
\usage{
old2NewSDMmodel(model, fold = NULL)
}
\arguments{
\item{model}{old SDMmodel object.}

\item{fold}{logical, used internally, don't use it directly.}
}
\value{
An \code{\linkS4class{SDMmodel}} object in the new format.
}
\description{
This is an help function to convert SDMmodel objects created with an old
version of \pkg{SDMtune} (v <= 0.2.0) into the new format.
}
\author{
Sergio Vignali
}
