% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help-functions.R
\name{old2NewSDMmodelCV}
\alias{old2NewSDMmodelCV}
\title{Convert Old SDMmodelCV object}
\usage{
old2NewSDMmodelCV(model)
}
\arguments{
\item{model}{old SDMmodelCV object.}
}
\value{
An \code{\linkS4class{SDMmodelCV}} object in the new format.
}
\description{
This is an help function to convert SDMmodelCV objects created with an old
version of \pkg{SDMtune} (v <= 0.2.0) into the new format.
}
\author{
Sergio Vignali
}
