% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help-functions.R
\name{old2NewSDMtune}
\alias{old2NewSDMtune}
\title{Convert Old SDMtune object}
\usage{
old2NewSDMtune(object)
}
\arguments{
\item{object}{old SDMtune object.}
}
\value{
An \code{\linkS4class{SDMtune}} object in the new format.
}
\description{
This is an help function to convert SDMtune objects created with an old
version of \pkg{SDMtune} (v <= 0.2.0) into the new format.
}
\author{
Sergio Vignali
}
