% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{old2NewSWD}
\alias{old2NewSWD}
\title{Convert Old SWD object}
\usage{
old2NewSWD(object1, object2 = NULL, fold = NULL)
}
\arguments{
\item{object1}{Old SWD object with presence or test locations.}

\item{object2}{Old SWD object with absence/background locations.}

\item{fold}{logical, used internally, don't use it directly.}
}
\value{
An \code{\linkS4class{SWD}} object in the new format.
}
\description{
This is an help function to convert SWD objects created with an old version
of \pkg{SDMtune} (v <= 0.2.0) into the new format.
}
\author{
Sergio Vignali
}
