% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{one_hot}
\alias{one_hot}
\title{One Hot Encoding}
\usage{
one_hot(x, levels)
}
\arguments{
\item{x}{factor. Factor to be encoded}

\item{levels}{character. Levels to be encoded.}
}
\value{
matrix with the One Hot encoded values
}
\description{
Converts a factor in One Hot matrix
}
\examples{
one_hot(1:10, c("1", "7"))
}
\author{
Sergio Vignali
}
