% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimiseModel1.R
\name{optimiseModel1}
\alias{optimiseModel1}
\title{Optimise Model}
\usage{
optimiseModel1(model, hypers, bg4test = NULL, test = NULL, pop = 20,
  gen = 5, metric = c("auc", "tss", "aicc"), env = NULL,
  parallel = FALSE, keep_best = 0.4, keep_random = 0.2,
  mutation_chance = 0.4, seed = NULL)
}
\arguments{
\item{model}{\link{SDMmodel} or \link{SDMmodelCV} object.}

\item{hypers}{named list containing the values of the hyperpatameters that
should be tuned, see details.}

\item{bg4test}{\link{SWD} object or NULL. Background locations used to get
subsamples, used only if bg hyperparameter is tuned, default is NULL.}

\item{test}{\link{SWD} object. Test dataset used to evaluate the model, not
used with aicc and \link{SDMmodelCV} objects, default is NULL.}

\item{pop}{numeric. Size of the population, default is 20.}

\item{gen}{numeric. Number of generations, default is 20.}

\item{metric}{character. The metric used to evaluate the models, possible
values are: "auc", "tss" and "aicc", default is "auc".}

\item{env}{\link{stack} containing the environmental variables, used only
with "aicc", default is NULL.}

\item{parallel}{logical, if TRUE it uses parallel computation, deafult is
FALSE.}

\item{keep_best}{numeric. Percentage of the best models in the population to
be retained during each iteration, expressed as decimal number. Default
is 0.4.}

\item{keep_random}{numeric. Probability of retaining the excluded models
during each iteration, expressed as decimal number. Default is 0.2.}

\item{mutation_chance}{numeric. Probability of mutation of the child models,
expressed as decimal number. Default is 0.4.}

\item{seed}{numeric. The value used to set the seed to have consistent
results, default is NULL.}
}
\value{
\link{SDMtune} object.
}
\description{
The function can be used to run a random search or a genetic algorithm to
optimise the model performance.
}
\details{
To know which hyperparameters can be tune you can use the output of
the function \link{get_tunable_args}.
}
\examples{
\dontrun{output <- optimiseModel(my_model, hypers = list( "reg" =
c(0.5, 1, 1.5), fcs = c("lq", "lqp", "lqph"), bgs = c(5000, 10000, 15000)),
bg4test = bg, test = my_val, pop = 20, gen = 10, seed = 25)}

}
\author{
Sergio Vignali
}
