% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizeModel.R
\name{optimizeModel}
\alias{optimizeModel}
\title{Optimize Model}
\usage{
optimizeModel(model, hypers, metric, test = NULL, bg4test = NULL,
  pop = 20, gen = 5, env = NULL, parallel = FALSE,
  keep_best = 0.4, keep_random = 0.2, mutation_chance = 0.4,
  seed = NULL)
}
\arguments{
\item{model}{\link{SDMmodel} or \link{SDMmodelCV} object.}

\item{hypers}{named list containing the values of the hyperparameters that
should be tuned, see details.}

\item{metric}{character. The metric used to evaluate the models, possible
values are: "auc", "tss" and "aicc".}

\item{test}{\link{SWD} object. Test dataset used to evaluate the model, not
used with aicc and \link{SDMmodelCV} objects, default is NULL.}

\item{bg4test}{\link{SWD} object or NULL. Background locations used to get
subsamples if the \strong{a} hyperparameter is tuned, default is NULL.}

\item{pop}{numeric. Size of the population, default is 20.}

\item{gen}{numeric. Number of generations, default is 20.}

\item{env}{\link{stack} containing the environmental variables, used only
with "aicc", default is NULL.}

\item{parallel}{logical, if TRUE it uses parallel computation, default is
FALSE.}

\item{keep_best}{numeric. Percentage of the best models in the population to
be retained during each iteration, expressed as decimal number. Default
is 0.4.}

\item{keep_random}{numeric. Probability of retaining the excluded models
during each iteration, expressed as decimal number. Default is 0.2.}

\item{mutation_chance}{numeric. Probability of mutation of the child models,
expressed as decimal number. Default is 0.4.}

\item{seed}{numeric. The value used to set the seed to have consistent
results, default is NULL.}
}
\value{
\link{SDMtune} object.
}
\description{
The function uses a Genetic Algorithm implementation to optimize the model
hyperparameter configuration according to the chosen metric.
}
\details{
To know which hyperparameters can be tuned you can use the output of
the function \link{get_tunable_args}. Parallel computation increases the
speed only for large datasets due to the time necessary to create the
cluster. Part of the code is inspired by
\href{https://blog.coast.ai/lets-evolve-a-neural-network-with-a-genetic
algorithm-code-included-8809bece164}{this post}
}
\examples{
\dontrun{output <- optimizeModel(my_model, hypers = list( "reg" =
c(0.5, 1, 1.5), "fc" = c("lq", "lqp", "lqph"), "a" = c(5000, 10000, 15000)),
bg4test = bg, test = my_val, pop = 20, gen = 10, seed = 25)}

}
\author{
Sergio Vignali
}
