% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDMmodel_class.R
\docType{methods}
\name{plot,SDMmodel,missing-method}
\alias{plot,SDMmodel,missing-method}
\title{Plot SDMmodel object}
\usage{
\S4method{plot}{SDMmodel,missing}(x)
}
\description{
Used only for the objects containing NN model. It plots the auc values for
all the epoch if monitored during training.
}
\author{
Sergio Vignali
}
