% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCor.R
\name{plotCor}
\alias{plotCor}
\title{Plot Correlation}
\usage{
plotCor(bg, method = "spearman", cor_th = NULL)
}
\arguments{
\item{bg}{\link{SWD} object used to compute the correlation matrix.}

\item{method}{character. The method used to compute the correlation matrix,
default is "spearman".}

\item{cor_th}{numeric. If provided it prints only the coefficients that are
higher or lower than the given threshold, default is NULL.}
}
\value{
The plot object.
}
\description{
Plot a correlation matrix heat map with the value of the correlation
coefficients according with the given method. If cor_th is passed then it
prints only the coefficients that are higher or lower than the given
threshold.
}
\details{
The code is inspired by the \href{www.sthda.com/english/wiki/ggplot2-quick-correlation-matrix-heatmap-r-software-and-data-visualization}{STHDA}
web page.
}
\examples{
\dontrun{
my_plot <- plotCorrelation(bgs, method = "spearman", cor_th = 0.7)}

}
\author{
Sergio Vignali
}
