% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotJk.R
\name{plotJk}
\alias{plotJk}
\title{Plot Jackknife Test}
\usage{
plotJk(model, jk, type = c("train", "test"))
}
\arguments{
\item{model}{Maxent object.}

\item{jk}{data.frame with the output of the doJk function.}

\item{type}{character, train or test to plot the train or test AUC values.}
}
\value{
The ggplot object.
}
\description{
Plot the Jackknife Test for variable importance.
}
\examples{
\dontrun{
plotJk(model, jk_test = 'AUC')}

}
\author{
Sergio Vignali
}
