% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPA.R
\name{plotPA}
\alias{plotPA}
\title{Plot Presence Absence Map}
\usage{
plotPA(map, th, colors = NULL, hr = FALSE, filename = NULL,
  format = "GTiff", ...)
}
\arguments{
\item{map}{\link{raster} object with the prediction.}

\item{th}{numeric. The threshold used to convert the output in a
presence/absence map.}

\item{colors}{vector. Colors to be used, default is NULL and uses red and
blue.}

\item{hr}{logical, if TRUE produces an output with high resolution, default
is FALSE.}

\item{filename}{character, if provided the raster map is saved in a file,
default is NULL.}

\item{format}{character. The output format, see \link{writeRaster} for all
the options, default is Geotiff.}

\item{...}{Additional arguments, see \link{writeRaster} for all the options.}
}
\value{
A ggplot object.
}
\description{
Plot a presence absence map using the given threshold.
}
\examples{
\dontrun{
plotPA(map, th = 7, colors = c("#7fbf7b", "#af8dc3"), filename = 'my_pa_map',
format = 'GTiff', overwrite = TRUE)}

}
\author{
Sergio Vignali
}
