% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotROC.R
\name{plotROC}
\alias{plotROC}
\title{Plot ROC curve}
\usage{
plotROC(model, val = NULL, test = NULL)
}
\arguments{
\item{model}{SDMmodel object.}

\item{val}{SWD the validation dataset, default is NULL.}

\item{test}{SWD the test dataset, default is NULL.}
}
\value{
The plot object.
}
\description{
Plot the ROC curve of the given model and print the AUC value.
}
\examples{
\dontrun{
plotROC(model, test)}

}
\author{
Sergio Vignali
}
