% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVarImp.R
\name{plotVarImp}
\alias{plotVarImp}
\title{Plot Varible Importance}
\usage{
plotVarImp(model, type = c("contribution", "permutation"), color = "grey")
}
\arguments{
\item{model}{Maxent object.}

\item{type}{character, contribution or permutation (or an abbreviation of them) are possible values.}

\item{color}{character. The color of the bar plot, default is grey.}
}
\value{
The ggplot object.
}
\description{
Plot the variable importance as a bar plot.
}
\examples{
\dontrun{
plotContribution(model, color = '#a1d99b')}

}
\author{
Sergio Vignali
}
