% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-ANN.R
\name{predict,ANN-method}
\alias{predict,ANN-method}
\title{Predict ANN}
\usage{
\S4method{predict}{ANN}(object, data, type, clamp)
}
\arguments{
\item{object}{\linkS4class{ANN} object.}

\item{data}{data.frame with the data for the prediction.}

\item{type}{Not used.}

\item{clamp}{Not used.}
}
\value{
A vector with the predicted values.
}
\description{
Predict the output for a new dataset from a trained ANN model.
}
\details{
Used by the \link{predict,SDMmodel-method}, not exported.
}
\author{
Sergio Vignali
}
