% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-BRT.R
\name{predict,BRT-method}
\alias{predict,BRT-method}
\title{Predict BRT}
\usage{
\S4method{predict}{BRT}(object, data, type, clamp)
}
\arguments{
\item{object}{\linkS4class{BRT} object.}

\item{data}{data.frame with the data for the prediction.}

\item{type}{Not used.}

\item{clamp}{Not used.}
}
\value{
A vector with the predicted values.
}
\description{
Predict the output for a new dataset from a trained BRT model.
}
\details{
Used by the \link{predict,SDMmodel-method}, not exported.

The function uses the number of tree defined to train the model and the
"response" type output.
}
\author{
Sergio Vignali
}
