% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-Maxnet.R
\name{predict,Maxnet-method}
\alias{predict,Maxnet-method}
\title{Predict Maxnet}
\usage{
\S4method{predict}{Maxnet}(
  object,
  data,
  type = c("link", "exponential", "cloglog", "logistic"),
  clamp = TRUE
)
}
\arguments{
\item{object}{\linkS4class{Maxnet} object.}

\item{data}{data.frame with the data for the prediction.}

\item{type}{character. Maxnet output type, possible values are "link",
"exponential", "cloglog" and "logistic".}

\item{clamp}{logical for clumping during prediction.}
}
\value{
A vector with the predicted values.
}
\description{
Predict the output for a new dataset from a trained Maxnet model.
}
\details{
Used by the \link{predict,SDMmodel-method}, not exported.
}
\author{
Sergio Vignali
}
