% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_NN.R
\docType{methods}
\name{predict,NN-method}
\alias{predict,NN-method}
\title{Predict NN}
\usage{
\S4method{predict}{NN}(object, data, type = "logistic", clamp = TRUE)
}
\arguments{
\item{object}{NN object.}

\item{data}{data.frame containing values used for the prediction.}

\item{type}{character. Output type, at the moment only "logistic is
available.}

\item{clamp}{logical for clumping during prediction, default is TRUE.}
}
\value{
A vector with the prediction
}
\description{
Predict output for a new dataset from a trained Neural Network model.
}
\examples{
\dontrun{
predict(model, my_dataset)}

}
\author{
Sergio Vignali
}
