% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-RF.R
\name{predict,RF-method}
\alias{predict,RF-method}
\title{Predict RF}
\usage{
\S4method{predict}{RF}(object, data, type, clamp)
}
\arguments{
\item{object}{\linkS4class{RF} object.}

\item{data}{data.frame with the data for the prediction.}

\item{type}{Not used.}

\item{clamp}{Not used.}
}
\value{
A vector with the predicted probabilities of class 1.
}
\description{
Predict the output for a new dataset from a trained RF model.
}
\details{
Used by the \link{predict,SDMmodel-method}, not exported.
}
\author{
Sergio Vignali
}
