% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.Maxent}
\alias{predict.Maxent}
\title{Predict Maxent}
\usage{
predict.Maxent(model, data, clamp = TRUE, parallel = FALSE,
  output_format = c("logistic", "cloglog", "raw"))
}
\arguments{
\item{model}{A Maxent model object.}

\item{data}{A data frame or a raster Stack.}

\item{clamp}{Flag for doing clumping during prediction, default is TRUE.}

\item{parallel}{Flag to use parallel computation during prediction, default is FALSE.}

\item{output_format}{The format of the output, possible values are "raw", "logistic" or "cloglog".}
}
\value{
A vector of prediction or a Raster object if data is a raster stack.
}
\description{
Predict Maxent
}
\details{
You need package to use parallel computation.
}
\examples{
\dontrun{
predic(model, predictors, output_format = "cloglog", parallel = TRUE)}

}
\author{
Sergio Vignali
}
