% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.Maxent.R
\docType{methods}
\name{predict,Maxent-method}
\alias{predict,Maxent-method}
\title{Predict Maxent}
\usage{
\S4method{predict}{Maxent}(object, data, clamp = TRUE, filename = "",
  format = "GTiff", extent = NULL, parallel = FALSE, progress = "", ...)
}
\arguments{
\item{object}{Maxent object.}

\item{data}{data.frame, \code{\link{SWD}}, \code{\link{stack}} or \code{\link{brick}}.}

\item{clamp}{logical for clumping during prediction, default is TRUE.}

\item{filename}{character. Output file name for the prediction map, if provided the output is
saved in a file.}

\item{format}{character. The output format, see \code{\link{writeRaster}} for all the options, default is "GTiff".}

\item{extent}{\link{Extent} object, if provided it restricts the prediction to the given
extent, default is NULL.}

\item{parallel}{logical to use parallel computation during prediction, default is FALSE.}

\item{progress}{character to display a progress bar: "text", "window" or "" (default) for no progress bar.}

\item{...}{Additional parameter to pass to the \code{\link{writeRaster}} function.}
}
\value{
A vector of prediction or a Raster object if data is a raster stack/brick.
}
\description{
Predict Maxent
}
\details{
You need package \link{snow} to use parallel computation and \code{\link{rgdal}} to save the prediction in a raster file.
}
\examples{
\dontrun{
predict(model, predictors, parallel = TRUE)}

}
\author{
Sergio Vignali
}
