% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareSWD.R
\name{prepareSWD}
\alias{prepareSWD}
\title{Prepare a SWD data set}
\usage{
prepareSWD(species, coords, env, categorical = NULL)
}
\arguments{
\item{species}{character. The name of the species.}

\item{coords}{data.frame. The coordinates of the presence or
absence/background locations.}

\item{env}{\link{stack} containing the environmental variables used to
extract the values at coordinate locations.}

\item{categorical}{vector indicating which of the environmental variable are
categorical, default is NULL.}
}
\value{
A \link{SWD} object
}
\description{
Given the coordinates, the species' name and the environmental variables,
the function creates a \link{SWD} object (sample with data).
}
\examples{
\dontrun{
swd <- prepareSWD(species = "Vultur gryphus", coords = p, env = env)}

}
\author{
Sergio Vignali
}
