% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{randomFolds}
\alias{randomFolds}
\title{Create Random Folds}
\usage{
randomFolds(data, k, only_presence = FALSE, seed = NULL)
}
\arguments{
\item{data}{\code{\linkS4class{SWD}} object that will be used to train the
model.}

\item{k}{integer. Number of fold used to create the partition.}

\item{only_presence}{logical, if \code{TRUE} the random folds are created
only for the presence locations, used manly for presence-only methods.}

\item{seed}{integer. The value used to set the seed for the fold partition,
default is \code{NULL}.}
}
\value{
matrix with each column representing a fold. with \code{TRUE} for
train locations and \code{FALSE} for test locations.
}
\description{
Create random folds for cross validation
}
\author{
Sergio Vignali
}
