% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomSearch.R
\name{randomSearch}
\alias{randomSearch}
\title{Random Search}
\usage{
randomSearch(model, hypers, metric, test = NULL, bg4test = NULL,
  pop = 20, env = NULL, parallel = FALSE, include_sm = FALSE,
  seed = NULL)
}
\arguments{
\item{model}{\link{SDMmodel} or \link{SDMmodelCV} object.}

\item{hypers}{named list containing the values of the hyperpatameters that
should be tuned, see details.}

\item{metric}{character. The metric used to evaluate the models, possible
values are: "auc", "tss" and "aicc".}

\item{test}{\link{SWD} object. Test dataset used to evaluate the model, not
used with aicc and \link{SDMmodelCV} objects, default is NULL.}

\item{bg4test}{\link{SWD} object or NULL. Background locations used to get
subsamples if \strong{a} hyperparameter is tuned, default is NULL.}

\item{pop}{numeric. Size of the population, default is 20.}

\item{env}{\link{stack} containing the environmental variables, used only
with "aicc", default is NULL.}

\item{parallel}{logical, if TRUE it uses parallel computation, deafult is
FALSE.}

\item{include_sm}{logical, include starting model. If TRUE the starting
model is included in the random population, default is FALSE.}

\item{seed}{numeric. The value used to set the seed to have consistent
results, default is NULL.}
}
\value{
\link{SDMtune} object.
}
\description{
The function performs a random search in the hyperparameters space, creating
a population of random models each one with a random combination of the
provided hyperparameters values.
}
\details{
To know which hyperparameters can be tune you can use the output of
the function \link{get_tunable_args}.
}
\examples{
\dontrun{output <- optimiseModel(my_model, hypers = list( "reg" =
c(0.5, 1, 1.5), "fc" = c("lq", "lqp", "lqph"), "a" = c(5000, 10000, 15000)),
bg4test = bg, test = my_val, pop = 30, seed = 25)}

}
\author{
Sergio Vignali
}
