% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NN_utils.R
\name{reshape_input}
\alias{reshape_input}
\title{Reshape Input}
\usage{
reshape_input(model, input_units)
}
\arguments{
\item{model}{NN object.}

\item{input_units}{numeric. The number of unit to be set in the new model.}
}
\value{
NN object with the new number of input units.
}
\description{
Reshape input units of a NN model
}
\examples{
\dontrun{reshape_input(model, 7)}

}
\author{
Sergio Vignali
}
