% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\name{response}
\alias{response}
\title{Response Curve}
\usage{
response(model, variable, maxent_output = c("cloglog", "logistic", "raw"),
  clamp = TRUE, marginal = FALSE, rug = FALSE, color = "red")
}
\arguments{
\item{model}{A MaxentModel object.}

\item{variable}{A variable name, used to plot only one variable.}

\item{marginal}{Flag to plot the marginal response courves, default is FALSE.}

\item{rug}{Flag to add the rug plot for presence locations, available only for continuous variables,
default is FALSE.}

\item{color}{The color of the curve, default is "red".}
}
\value{
The plot objects
}
\description{
Plot the Responce Courve of the given environmental variable
}
\examples{
\dontrun{responce(model, variable = 'bio12', marginal = TRUE, rug = TRUE)}

}
\author{
Sergio Vignali
}
