% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMaxent.R
\name{runMaxent}
\alias{runMaxent}
\title{Run MaxEnt}
\usage{
runMaxent(train, bg, rm, fc, test = NULL, output_format = "cloglog",
  response_curves = FALSE, iterations = 500, extra_args = NULL,
  folder = NULL)
}
\arguments{
\item{train}{The data frame used to train the model given as SWD object.}

\item{bg}{The data frame with the background locations given as SWD object.}

\item{rm}{The value of the regularization multiplier.}

\item{fc}{The value of the feature combination, possible values are combinations of
"L", "Q", "P", "H" and "T".}

\item{test}{The test dataset given as SWD object, default is NULL.}

\item{output_format}{The MaxEnt output format, possible values are "Logistic", "Cloglog",
"Cumulative" and "Raw", default value "Cloglog".}

\item{response_curves}{Flag to compute the response curves, default is FALSE.}

\item{iterations}{Number of iterations used by the Maxent alghoritm, default is 500.}

\item{extra_args}{Extra arguments used to run MaxEnt, e.g. "removeduplicates=false", default
is NULL.}

\item{folder}{The folder name where to save the MaxEnt output, default is NULL meaning a
temporary folder. The folder is created in the working directory.}

\item{jackknife}{Flag to activate the jackknife test, default FALSE.}

\item{threads}{Number of threads used by MaxEnt, default is 1.}
}
\value{
The output of MaxEnt as Maxent object.
}
\description{
Run a MaxEnt model using the dismo package.
}
\examples{
\dontrun{model <- runMaxent(train, bg, rm, response_curves = TRUE))}

}
\author{
Sergio Vignali
}
