% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDMtuneChart.R
\name{saveChart}
\alias{saveChart}
\title{Save Chart}
\usage{
saveChart(x, filename)
}
\arguments{
\item{x}{the output of the \link{plot,SDMtune,missing-method}.}

\item{filename}{character containing the name of the file in which to save
the chart.}
}
\description{
Save the interactive chart in a html file
}
\details{
Note that the \code{filemane} argument should be provided without
the extension (i.e. "my_chart" and not "my_chart.html").
}
\examples{
\donttest{
# Acquire environmental variables
files <- list.files(path = file.path(system.file(package = "dismo"), "ex"),
                    pattern = "grd", full.names = TRUE)
predictors <- raster::stack(files)

# Prepare presence locations
p_coords <- condor[, 1:2]

# Prepare background locations
bg_coords <- dismo::randomPoints(predictors, 5000)

# Create SWD object
presence <- prepareSWD(species = "Vultur gryphus", coords = p_coords,
                       env = predictors, categorical = "biome")
bg <- prepareSWD(species = "Vultur gryphus", coords = bg_coords,
                 env = predictors, categorical = "biome")

# Split presence locations in training (80\%) and testing (20\%) datasets
datasets <- trainValTest(presence, test = 0.2)
train <- datasets[[1]]
test <- datasets[[2]]

# Train a model
model <- train(method = "Maxnet", p = train, a = bg, fc = "l")

# Define the hyperparameters to test
h <- list(reg = 1:2, fc = c("lqp", "lqph"), a = c(1000, 2000))

# Run the gridSearch function using as metric the AUC
output <- gridSearch(model, hypers = h, metric = "auc", test = test,
                     bg4test = bg)

# Plot the output
plot(output, title = "My experiment")

# Plot the interactive chart
p <- plot(output, title = "My experiment", interactive = TRUE)
p
# Save the chart in a html file
saveChart(p, "my_chart")
}
}
\seealso{
\link{plot,SDMtune,missing-method}
}
\author{
Sergio Vignali
}
