% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NN_utils.R
\name{saveNN}
\alias{saveNN}
\title{Save NN}
\usage{
saveNN(model, file_name = NULL)
}
\arguments{
\item{model}{SDMmodel object.}

\item{file_name}{character. The name of the file in which to save the model,
default is NULL and it uses the name of the species.}
}
\description{
Save a SDMmodel object containing a NN model.
}
\details{
The function creates two files, a hdf5 file for the serializes keras
object and a Rds file for the SDMmodel object. The file are saved in the
working directory.
}
\examples{
\dontrun{saveNN(model, "myNN")}

}
\author{
Sergio Vignali
}
