% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDMmodel_class.R
\docType{class}
\name{SDMmodel-class}
\alias{SDMmodel-class}
\alias{SDMmodel}
\title{SDMmodel}
\description{
This Class represents a MaxEnt model objects and hosts all the information related to the model.
}
\section{Slots}{

\describe{
\item{\code{presence}}{SWD. The presence locations used to train the model.}

\item{\code{background}}{SWD. The backgorund locations used to train the model.}

\item{\code{model}}{Maxent or Maxnet object.}
}}

\author{
Sergio Vignali
}
