% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDMmodel2MaxEnt.R
\name{SDMmodel2MaxEnt}
\alias{SDMmodel2MaxEnt}
\title{SDMmodel2MaxEnt}
\usage{
SDMmodel2MaxEnt(model)
}
\arguments{
\item{model}{\link{SDMmodel} object to be converted.}
}
\value{
The converted \link{MaxEnt} object.
}
\description{
Converts a \link{SDMmodel} object containing a \link{Maxent} model to a dismo
\link{MaxEnt} object.
}
\examples{
\dontrun{SDMmodel2MaxEnt(model)}

}
\author{
Sergio Vignali
}
