% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDMsel_class.R
\docType{class}
\name{SDMsel-class}
\alias{SDMsel-class}
\alias{SDMsel}
\title{SDMsel class}
\description{
SDMsel class
}
\section{Slots}{

\describe{
\item{\code{results}}{data.frame. Results with the evaluation of the models.}

\item{\code{models}}{list. List of \link{Maxent} objects}
}}

\author{
Sergio Vignali
}
