% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDMsel.R
\docType{package}
\name{SDMsel}
\alias{SDMsel}
\alias{SDMsel-package}
\title{SDMsel: A package for tuning Species Distribution Models.}
\description{
\strong{SDMsel} (Species Distribution Model selection) provides a framework that
facilitates users in preparing data for analysis, train and evaluate models.
It also implements functions for data driven variable selection and model
tuning and includes some utilities to display results (at the moment it
implements MaxEnt and Maxnet models). SDMsel uses its own script to predict
MaxEnt models that results to be much faster for large datasets than native
prediction made using Java software. This reduce considerably the computation
time when tuninig the model using the AICc.
}
