% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDMtune_class.R
\docType{class}
\name{SDMtune-class}
\alias{SDMtune-class}
\alias{SDMtune}
\title{SDMtune class}
\description{
Class used to save the results of one of the following functions:
\link{gridSearch}, \link{randomSearch} or \link{optimizeModel}.
}
\section{Slots}{

\describe{
\item{\code{results}}{data.frame. Results with the evaluation of the models.}

\item{\code{models}}{list. List of \link{Maxent} objects}
}}

\author{
Sergio Vignali
}
