% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDMtune-package.R
\docType{package}
\name{SDMtune-package}
\alias{SDMtune-package}
\alias{_PACKAGE}
\title{SDMtune: Species Distribution Model Selection}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

User-friendly framework that enables the training and the evaluation of species distribution models (SDMs). The package implements functions for data driven variable selection and model tuning and includes numerous utilities to display the results. All the functions used to select variables or to tune model hyperparameters have an interactive real-time chart displayed in the 'RStudio' viewer pane during their execution.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://consbiol-unibern.github.io/SDMtune/}
  \item \url{https://github.com/ConsBiol-unibern/SDMtune}
  \item Report bugs at \url{https://github.com/ConsBiol-unibern/SDMtune/issues}
}

}
\author{
\strong{Maintainer}: Sergio Vignali \email{sergio.vignali@iee.unibe.ch} (\href{https://orcid.org/0000-0002-3390-5442}{ORCID})

Authors:
\itemize{
  \item Arnaud Barras (\href{https://orcid.org/0000-0003-0850-6965}{ORCID})
  \item Veronika Braunisch (\href{https://orcid.org/0000-0001-7035-4662}{ORCID})
}

Other contributors:
\itemize{
  \item Conservation Biology - University of Bern [funder]
}

}
\keyword{internal}
