% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDMtune.R
\docType{package}
\name{SDMtune-pkg}
\alias{SDMtune-pkg}
\alias{SDMtune-pkg-package}
\title{SDMtune: A package for tuning Species Distribution Models.}
\description{
\strong{SDMtune} provides a framework that facilitates users in preparing data for
analysis, train and evaluate SDMs. It also implements functions for data
driven variable selection and model tuning and includes some utilities to
display results (at the moment only the Maximum Entropy method is available
using the Java implementation through the "dismo" package and the R
implementation through the "maxnet" package). All the functions used to
select variables or to tune model hyperparameters have an interactive
real-time chart that is displayed in the RStudio viewer pane during the
function execution. SDMtune uses its own script to predict MaxEnt models that
results to be much faster for large datasets than native prediction made
using the Java software. This reduces considerably the computation time when
tuning the model using the AICc.
}
