% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWD_class.R
\docType{class}
\name{SWD-class}
\alias{SWD-class}
\alias{SWD}
\title{Sample With Data}
\description{
Object similar to the MaxEnt SWD format that hosts the species name, the
coordinates of the locations and the value of the environmental variables at
the location places.
}
\details{
\code{data} and \code{coords} must have the same number of rows;
\code{pa} must have length equal to \code{row(data)}.
}
\section{Slots}{

\describe{
\item{\code{species}}{character. Name of the species.}

\item{\code{coords}}{data.frame. Coordinates of the locations.}

\item{\code{data}}{data.frame. Value of the environmental variables at location
sites.}

\item{\code{pa}}{numeric. Vector with \code{1} for presence and \code{0} for
absence/background locations.}
}}

\author{
Sergio Vignali
}
