% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swd2csv.R
\name{swd2csv}
\alias{swd2csv}
\title{SWD to csv}
\usage{
swd2csv(swd, file_name)
}
\arguments{
\item{swd}{A MaxentSWD object.}

\item{file_name}{The name of the file where to save the object.}
}
\description{
Save a MaxentSWD object in a csv file.
}
\examples{
\dontrun{
swd2csv(train, "train_data.csv")}

}
\author{
Sergio Vignali
}
