% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thinData.R
\name{thinData}
\alias{thinData}
\title{Thin Data}
\usage{
thinData(coords, env)
}
\arguments{
\item{coords}{data.frame or matrix with the coordinates.}

\item{env}{\link{stack} or \link{brick} containing the environmental variables,
or a single raster layer.}
}
\value{
a data frame with the thinned locations.
}
\description{
Remove all but one location per raster cell. The function selects randomly one
location if more than one fall within the same raster cell.
}
\details{
\strong{coords} and \strong{env} must have the same coordinate reference system.
}
\examples{
\dontrun{
thinData(my_coords, my_env)}

}
\author{
Sergio Vignali
}
