% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{train}
\alias{train}
\title{Train}
\usage{
train(method = c("Maxent", "Maxnet"), presence, bg, replicates = 1,
  ...)
}
\arguments{
\item{method}{character. Possible values are "Maxent" or "Maxnet".}

\item{presence}{SWD object with the presence locations.}

\item{bg}{SWD object with the background locations.}

\item{replicates}{numeric. Number of replicates, used for cross validation.
Default is 1, meaning no cross validation is performed.}

\item{...}{Arguments passed to the relative functions, see \link{trainMaxent}
or \link{trainMaxnet} for details related to the different methods.}
}
\value{
A SWDmodel or SDMmodelCV object.
}
\description{
Train a model using the given method.
}
\examples{
\dontrun{
model <- train("Maxnet", presence, bg, reg = 2, fc = "lqp")}

}
\author{
Sergio Vignali
}
