% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainMaxnet.R
\name{trainMaxnet}
\alias{trainMaxnet}
\title{Train Maxnet model}
\usage{
trainMaxnet(presence, bg, rm, fc, type = c("cloglog", "logistic", "raw"),
  test = NULL)
}
\arguments{
\item{presence}{SWD object with the presence locations.}

\item{bg}{SWD object with the background locations.}

\item{rm}{numeric. The value of the regularization multiplier.}

\item{type}{The MaxEnt output type, possible values are "Cloglog", "Logistic",
and "Raw", default value "Cloglog".}

\item{test}{SWD object with the test locations, default is NULL.}
}
\value{
Maxent object.
}
\description{
Train a Maxnet model using the maxnet package.
}
\examples{
\dontrun{model <- trainMaxnet(presence, bg, rm)}

}
\author{
Sergio Vignali
}
