% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainValTest.R
\name{trainValTest}
\alias{trainValTest}
\title{Train, Validation and Test datasets}
\usage{
trainValTest(df, val, test, seed = NULL)
}
\arguments{
\item{df}{data.frame containing the data that have to be splitted in train, validation and test datasets.}

\item{val}{numeric. The percentage of data withhold for validation.}

\item{test}{numeric. The percentage of data withhold for testing.}

\item{seed}{numeric. The value used to set the seed in order to have consistent results, default is NULL.}
}
\value{
A list with the train, validation and test data sets or train and test sets accordingly.
}
\description{
Split a dataset randomly in train, validation and test datasets. If val parameter is set to zero then
the function returns only the train and the test datasets.
}
\examples{
\dontrun{
sets <- trainValTest(presence, val = 0.3, test = 0.1, set_seed = 25)}

}
\author{
Sergio Vignali
}
