% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuneBg.R
\name{tuneBg}
\alias{tuneBg}
\title{Tune number of Background locations}
\usage{
tuneBg(model, bg4test, bgs, metric = c("auc", "tss", "aicc"),
  test = NULL, env = NULL, parallel = FALSE, seed = NULL)
}
\arguments{
\item{model}{SDMmodel object.}

\item{bg4test}{SWD. The dataset with the maximum number of background
locations to be tested given as MaxentSWD object.}

\item{bgs}{vector. A sequence of number of background locations to test. The
maximum number must be lower than the maximum number of background locations
in bg4test.}

\item{metric}{character. The metric used to evaluate the models, possible
values are: "auc", "tss" and "aicc", default is "auc".}

\item{test}{SWD. Test dataset used to evaluate the model, not used with aicc
or SDMmodelCV objects, default is NULL.}

\item{env}{\link{stack} or \link{brick} containing the environmental
variables, used only with "aicc", default is NULL.}

\item{parallel}{logical, if TRUE it uses parallel computation, deafult is
FALSE. Used only with AICc.}

\item{seed}{integer. The value used to set the seed in order to have
consistent results, default is NULL.}
}
\value{
A \link{SDMtune} object.
}
\description{
Test different sizes of background locations.
}
\details{
You need package \pkg{snow} to use parallel computation and
\pkg{rgdal} to save the prediction in a raster file. Parallel computation
increases the speed only for big datasets due to the time necessary to create
the cluster.
}
\examples{
\dontrun{
bg_test <- tuneBg(model, bg4test,bgs = seq(5000, 30000, 5000),
metric = "auc", test = test, seed = 25)}

}
\seealso{
Other tuning functions: \code{\link{tuneFC}},
  \code{\link{tuneReg}}
}
\author{
Sergio Vignali
}
\concept{tuning functions}
