% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuneIt.R
\name{tuneIt}
\alias{tuneIt}
\title{Tune number of Iterations}
\usage{
tuneIt(model, its, metric = c("auc", "tss", "aicc"), env = NULL,
  parallel = FALSE)
}
\arguments{
\item{model}{Maxent object.}

\item{its}{vector. A sequence of number of iterations to be tested.}

\item{metric}{character. The metric used to evaluate the models, possible values are:
"auc", "tss" and "aicc", default is "auc".}

\item{env}{\link{stack} or \link{brick} containing the environmental variables,
used only with "aicc", default is NULL.}

\item{parallel}{logical, if TRUE it uses parallel computation, deafult is FALSE.}
}
\value{
A \link{SDMtune} object.
}
\description{
Test different number of iterations used by the optimization algorithm in MaxEnt.
}
\details{
You need package \pkg{snow} to use parallel computation and \pkg{rgdal}
to save the prediction in a raster file. Parallel computation increases the speed
only for big datasets due to the time necessary to create the cluster.
}
\examples{
\dontrun{
it_test <- tuneIt(model, its = seq(300, 900, 200), metric = "tss")}

}
\seealso{
Other tuning functions: \code{\link{tuneBg}},
  \code{\link{tuneFC}}, \code{\link{tuneRM}}
}
\author{
Sergio Vignali
}
