% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuneRM.R
\name{tuneRM}
\alias{tuneRM}
\title{Tune Regularization Multiplier}
\usage{
tuneRM(model, rms, metric = c("auc", "tss", "aicc"), env = NULL,
  parallel = FALSE, extra_args = NULL)
}
\arguments{
\item{model}{Maxent object.}

\item{rms}{vector. A sequence of regularization multipliers to test.}

\item{metric}{character. The metric used to evaluate the models, possible values are:
"auc", "tss" and "aicc", default is "auc".}

\item{env}{\link{stack} or \link{brick} containing the environmental variables,
used only with "aicc", default is NULL.}

\item{parallel}{logical, if TRUE it uses parallel computation, deafult is FALSE.}

\item{extra_args}{vector. Extra arguments used to run MaxEnt.}
}
\value{
A \link{SDMsel} object.
}
\description{
Given a sequence of regularization multipliers, the function runs several
MaxEnt models increasing the regularization multiplier.
}
\details{
You need package \pkg{snow} to use parallel computation and \pkg{rgdal}
to save the prediction in a raster file. Parallel computation increases the speed
only for big datasets due to the time necessary to create the cluster.
The minimum value of rm allow is 0.001, if lower MaxEnt crasches.
}
\examples{
\dontrun{
fc_test <- tuneRM(model, metric = "aicc", env = predictors, parallel = T)}

}
\author{
Sergio Vignali
}
