% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varImp.R
\name{varImp}
\alias{varImp}
\title{Variable Importance}
\usage{
varImp(model)
}
\arguments{
\item{model}{Maxent object.}
}
\value{
A data frame with the variable importance.
}
\description{
Shows the percent contribution and permutation importance of the variable used to train the model.
}
\examples{
\dontrun{
varContribution(model)}

}
\author{
Sergio Vignali
}
