% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varSel.R
\name{varSel}
\alias{varSel}
\title{Variable Selection}
\usage{
varSel(model, bg4cor, metric = c("auc", "tss", "aicc"), env = NULL,
  parallel = FALSE, rm = 0.001, method = "spearman", cor_th = 0.7,
  use_percent = FALSE)
}
\arguments{
\item{model}{Maxent object.}

\item{bg4cor}{SWD object. Background locations used to test the correlation between
environmental variables.}

\item{metric}{character. The metric used to evaluate the models, possible values are:
"auc", "tss" and "aicc", default is "auc".}

\item{env}{\link{stack} or \link{brick} containing the environmental variables,
used only with "aicc", default is NULL.}

\item{parallel}{logical, if TRUE it uses parallel computation, deafult is FALSE. Used only with AICc.}

\item{rm}{integer. The value of the regularization paramiter to use during computation,
default is 0.001, see details.}

\item{method}{character. The method used to comput the correlation matrix, default "spearman".}

\item{cor_th}{numeric. The correlation threshold used to select highly correlated variables,
default is 0.7.}

\item{use_percent}{logical, if TRUE it uses the percent contribution instead of the
permutation importance.}
}
\value{
The name of the selected variables.
}
\description{
The function performs a data-driven variable selection. Starting from the provided model
it iterates through all the variables starting from the one with the highest contribution
(percent or permutation). If the variable is correlated with other variables (according to
the given method and threshold) it performs a Jackknife test and among the correlated
variables it removes the one that results in the best performing model when removed
(according to the given metric). The process is repeated untill the remaining variables
are not highly correlated anymore.
}
\details{
You need package \pkg{snow} to use parallel computation. Parallel computation increases the speed
only for big datasets due to the time necessary to create the cluster.
We should write something more... I will refer to our paper for the explanations...
}
\examples{
\dontrun{
varSel(model, bg, use_permutation = T)}

}
\author{
Sergio Vignali
}
