% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varSelection.R
\name{varSelection}
\alias{varSelection}
\title{Variable Selection}
\usage{
varSelection(model, bg4cor, rm = 0.001, method = "spearman", cor_th = 0.7,
  use_permutation = TRUE)
}
\arguments{
\item{model}{A MaxentModel object.}

\item{bg4cor}{Background locations used to test the correlation between environmental variables,
given as MaxentSWD object.}

\item{rm}{integer. The value of the regularization paramiter to use during computation,
default is 0.001, see details.}

\item{method}{The method used to comput the correlation matrix, default "spearman".}

\item{cor_th}{The correlation threshold used to select highly correlated variables, default is 0.7.}

\item{use_permutation}{Flag to select the permutation importance or the percent contribution.}
}
\value{
The name of the selected variables.
}
\description{
The function performs a data driven variable selection. Starting from provided model
it iterates through the .......
}
\details{
Here we have to write something... I will refer to the paper for the explanations...
}
\examples{
\dontrun{
varSelection(model, bg, use_permutation = T)}

}
\author{
Sergio Vignali
}
