% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aicc.R
\name{aicc}
\alias{aicc}
\title{AICc}
\usage{
aicc(model, env, parallel = FALSE)
}
\arguments{
\item{model}{\link{SDMmodel} object.}

\item{env}{\link{stack} containing the environmental variables.}

\item{parallel}{logical, if TRUE it uses parallel computation, deafult is
FALSE.}
}
\value{
The computed AICc
}
\description{
Compute the Akaike Information Criterion corrected for small samples size
(Warren and Seifert 2011).
}
\details{
Parallel computation increases the speed only for big datasets due
to the time necessary to create the cluster.
}
\examples{
\dontrun{
aicc(model, predictors, parallel = T)}

}
\references{
Warren D.L., Seifert S.N., (2011). Ecological niche modeling in
Maxent: the importance of model complexity and the performance of model
selection criteria. Ecological Applications, 21(2), 335–342.
}
\author{
Sergio Vignali
}
