% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aicc.R
\name{aicc}
\alias{aicc}
\title{AICc}
\usage{
aicc(model, env, parallel = FALSE)
}
\arguments{
\item{model}{Maxent object.}

\item{env}{\link{stack} or \link{brick} containing the environmental variables.}

\item{parallel}{logical, if TRUE it uses parallel computation, deafult is FALSE.}
}
\value{
The computed AICc
}
\description{
Compute Akaike Information Criterion corrected for small samples size
\insertCite{Warren2011}{SDMsel}
}
\details{
You need package \pkg{snow} to use parallel computation. Parallel computation increases the speed
only for big datasets due to the time necessary to create the cluster.
}
\examples{
\dontrun{
aicc(model, predictors, parallel = T)}

}
\references{
\insertRef{Warren2011}{SDMsel}
}
\author{
Sergio Vignali
}
