% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aicc.R
\name{aicc}
\alias{aicc}
\title{AICc}
\usage{
aicc(model, env, parallel = FALSE)
}
\arguments{
\item{model}{\linkS4class{SDMmodel} object.}

\item{env}{\link[raster]{stack} containing the environmental
variables.}

\item{parallel}{deprecated.}
}
\value{
The computed AICc
}
\description{
Compute the Akaike Information Criterion corrected for small samples size
(Warren and Seifert, 2011).
}
\details{
The function is available only for \strong{Maxent} and \strong{Maxnet}
methods.
}
\examples{
# Acquire environmental variables
files <- list.files(path = file.path(system.file(package = "dismo"), "ex"),
                    pattern = "grd", full.names = TRUE)
predictors <- raster::stack(files)

# Prepare presence and background locations
p_coords <- virtualSp$presence
bg_coords <- virtualSp$background

# Create SWD object
data <- prepareSWD(species = "Virtual species", p = p_coords, a = bg_coords,
                   env = predictors, categorical = "biome")

# Train a model
model <- train(method = "Maxnet", data = data, fc = "l")

# Compute the AICc
aicc(model, predictors)
}
\references{
Warren D.L., Seifert S.N., (2011). Ecological niche modeling in
Maxent: the importance of model complexity and the performance of model
selection criteria. Ecological Applications, 21(2), 335–342.
}
\seealso{
\link{auc} and \link{tss}.
}
\author{
Sergio Vignali
}
