% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{auc}
\alias{auc}
\title{AUC}
\usage{
auc(model, test = NULL, a = NULL)
}
\arguments{
\item{model}{\link{SDMmodel} or \link{SDMmodelCV} object.}

\item{test}{\link{SWD} test locations for \link{SDMmodel} objects or logical
for \link{SDMmodelCV} objects, if not provided it computes the train AUC,
default is NULL.}

\item{a}{\link{SWD} absence or background locations used to compute the AUC
by the permutation importance function, default is NULL.}
}
\value{
The value of the AUC.
}
\description{
Compute the AUC using the Man-Whitney U Test formula.
}
\details{
If the model is a \link{SDMmodelCV} object, the function computes
the mean of the training or testing AUC values of the different replicates.
}
\examples{
\dontrun{
auc(my_model, test = test_dataset)}

}
\author{
Sergio Vignali
}
