% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{auc}
\alias{auc}
\title{AUC}
\usage{
auc(model, test = NULL, bg = NULL)
}
\arguments{
\item{model}{SDMmodel object.}

\item{test}{SWD test locations, if not provided it computes the train AUC,
default is NULL.}

\item{bg}{SWD backgroung locations used to compute the AUC
by the permutation importance function, defaul is NULL.}
}
\value{
The value of the AUC.
}
\description{
Compute the AUC using the Man-Whitney U Test formula
}
\examples{
\dontrun{
auc(my_model, test = test_dataset)}

}
\author{
Sergio Vignali
}
