% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{auc}
\alias{auc}
\title{AUC}
\usage{
auc(model, test = NULL, bg = NULL)
}
\arguments{
\item{model}{SDMmodel or SDMmodelCV object.}

\item{test}{SWD test locations for SDMmodel objects or TRUE for SDModelCV
objects, if not provided it computes the train AUC, default is NULL.}

\item{bg}{SWD backgroung locations used to compute the AUC
by the permutation importance function, defaul is NULL.}
}
\value{
The value of the AUC.
}
\description{
Compute the AUC using the Man-Whitney U Test formula.
}
\details{
If the model is a SDMmodelCV object, the function computes the mean
of the training or testing AUC values of the different replicates.
}
\examples{
\dontrun{
auc(my_model, test = test_dataset)}

}
\author{
Sergio Vignali
}
