% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRT-class.R
\docType{class}
\name{BRT-class}
\alias{BRT-class}
\alias{BRT}
\alias{show,BRT-method}
\title{Boosted Regression Tree}
\usage{
\S4method{show}{BRT}(object)
}
\arguments{
\item{object}{BRT object}
}
\description{
This Class represents a Boosted Regression Tree model objects and hosts all
the information related to the model.
}
\details{
See \link{gbm} for the meaning of the slots.
}
\section{Slots}{

\describe{
\item{\code{distribution}}{character. Name of the used distribution.}

\item{\code{n.trees}}{integer. Maximum number of grown trees.}

\item{\code{interaction.depth}}{integer. Maximum depth of each tree.}

\item{\code{shrinkage}}{numeric. The shrinkage parameter.}

\item{\code{bag.fraction}}{numeric. Random fraction of data used in the tree
expansion.}

\item{\code{model}}{\link{gbm}. The Boosted Regression Tree model object.}
}}

\author{
Sergio Vignali
}
