% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confMatrix.R
\name{confMatrix}
\alias{confMatrix}
\title{Confusion Matrix}
\usage{
confMatrix(model, presence = NULL, th = NULL)
}
\arguments{
\item{model}{Maxent object.}

\item{presence}{SWD presence locations, if not provided it uses the train dataset,
default is NULL.}

\item{th}{numeric if provided it computes the evaluation at the given threshold,
default is NULL.}
}
\value{
The Confusion Matrix for all the used thresholds.
}
\description{
Compute the Confusion Matrix for threshold values varying from 0 to 1
}
\examples{
\dontrun{
confMatrix(my_model, presence = test_dataset)}
}
