% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertFolds.R
\name{convertFolds}
\alias{convertFolds}
\title{Convert Folds}
\usage{
convertFolds(x, data)
}
\arguments{
\item{x}{list. Folds to be converted.}

\item{data}{\code{\linkS4class{SWD}} object used to create the folds.}
}
\value{
list with two matrices, the first for the training and the second for
the testing dataset. Each column of one matrix represents a fold with
\code{TRUE} for the locations included in and \code{FALSE} excluded from the
partition.
}
\description{
Convert folds created with other packages into a format valid for
\pkg{SDMtune}.
}
\details{
The function can convert folds created with the following packages:
ENMeval and CVblocks.
}
\examples{
\donttest{
# Acquire environmental variables
files <- list.files(path = file.path(system.file(package = "dismo"), "ex"),
                    pattern = "grd", full.names = TRUE)
predictors <- raster::stack(files)

# Prepare presence and background locations
p_coords <- virtualSp$presence
bg_coords <- virtualSp$background

data <- prepareSWD(species = "Virtual species", p = p_coords, a = bg_coords,
                   env = predictors, categorical = "biome")

# Example using the ENMeval package
library(ENMeval)
# Block partition
block_folds <- get.block(occ = data@data[data@pa == 1, ],
                         bg.coords = data@data[data@pa == 0, ])
folds <- convertFolds(block_folds)
# Checkerboard1 partition
cb_folds <- get.checkerboard1(occ = data@data[data@pa == 1, ],
                         env = predictors,
                         bg.coords = data@data[data@pa == 0, ],
                         aggregation.factor = 4)
folds <- convertFolds(cb_folds)
}
}
\author{
Sergio Vignali
}
