% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corVar.R
\name{corVar}
\alias{corVar}
\title{Print Correlated Variables}
\usage{
corVar(loc, method = "spearman", cor_th = 0.7)
}
\arguments{
\item{loc}{SWD. Locations used to test the correlation between environmental variables.}

\item{method}{character. The method used to comput the correlation matrix, default "spearman".}

\item{cor_th}{numeric. The correlation threshold used to select highly correlated variables, default is 0.7.}
}
\value{
The name of the correlated variables.
}
\description{
Utility that prints the name of correlated variables according with the provided
method and correlation threashold
}
\examples{
\dontrun{
corVar(loc, method = 'pearson', cor_th = 0.6)}

}
\author{
Sergio Vignali
}
