% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corVar.R
\name{corVar}
\alias{corVar}
\title{Print Correlated Variables}
\usage{
corVar(bg, method = "spearman", cor_th = NULL)
}
\arguments{
\item{bg}{\link{SWD}. Locations used to test the correlation between
environmental variables.}

\item{method}{character. The method used to compute the correlation matrix,
default "spearman".}

\item{cor_th}{numeric. If provided it prints only the variables whose
correlation coefficient is higer or lover than the given threshold, default
is NULL.}
}
\value{
The name of the correlated variables.
}
\description{
Utility that prints the name of correlated variables according with the
provided method and correlation threashold.
}
\examples{
\dontrun{
corVar(bg, method = 'pearson', cor_th = 0.6)}

}
\author{
Sergio Vignali
}
