% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doJk.R
\name{doJk}
\alias{doJk}
\title{Jackknife Test}
\usage{
doJk(model, metric, variables = NULL, test = NULL, with_only = TRUE,
  env = NULL, parallel = FALSE, return_models = FALSE)
}
\arguments{
\item{model}{\link{SDMmodel} or \link{SDMmodelCV} object.}

\item{metric}{character. The metric used to evaluate the models, possible
values are: "auc", "tss" and "aicc".}

\item{variables}{vector. Variables used for the test, if not provided it
takes all the variables used to train the model, default is NULL.}

\item{test}{\link{SWD}. If provided it reports the result also for the test
dataset. Not used for \strong{aicc} and \link{SDMmodelCV}.}

\item{with_only}{logical. If TRUE it runs the test also for each variable in
isolation, default is TRUE.}

\item{env}{\link{stack} containing the environmental variables, used only
with "aicc", default is NULL.}

\item{parallel}{logical, if TRUE it uses parallel computation, default is
FALSE. Used only with AICc.}

\item{return_models}{logical, If TRUE returns all the models together with
the test result, default is FALSE.}
}
\value{
A data frame with the test results. If \strong{return_model = TRUE} it
returns a list containing the test results together with the models.
}
\description{
Run the Jackknife test for variable importance removing one variable at time.
}
\details{
Parallel computation increases the speed only for large datasets due
to the time necessary to create the cluster.
}
\examples{
\dontrun{
doJk(model, variable = c('bio1', 'bio12'), with_only = TRUE)}

}
\author{
Sergio Vignali
}
