% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSubsample.R
\name{getSubsample}
\alias{getSubsample}
\title{Get Dataset Subsample}
\usage{
getSubsample(dataset, size, seed = NULL)
}
\arguments{
\item{dataset}{\link{SWD} object.}

\item{size}{numeric. The size of the sub sample.}

\item{seed}{numeric. The value used to set the seed in order to have
consistent results, default is NULL.}
}
\value{
The sub sample as \link{SWD} object.
}
\description{
Get a random subsample of an \link{SWD} object.
}
\examples{
\dontrun{
getSubsample(data, value = 4500)}

}
\author{
Sergio Vignali
}
