% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Maxent-class.R
\docType{class}
\name{Maxent-class}
\alias{Maxent-class}
\title{Maxent}
\description{
This Class represents a MaxEnt model objects and hosts all the information
related to the model.
}
\section{Slots}{

\describe{
\item{\code{results}}{matrix. The result that usually MaxEnt provide as a csv file.}

\item{\code{reg}}{numeric. The value of the regularization multiplier used to train
the model.}

\item{\code{fc}}{character. The feature class combination used to train the model.}

\item{\code{iter}}{numeric. The number of iterations used to train the model.}

\item{\code{extra_args}}{character. Extra arguments used to run MaxEnt.}

\item{\code{lambdas}}{vector. The lambdas parameters of the model.}

\item{\code{coeff}}{data.frame. The lambda coefficients of the model.}

\item{\code{formula}}{formula. The formula used to make prediction.}

\item{\code{lpn}}{numeric. Linear Predictor Normalizer.}

\item{\code{dn}}{numeric. Density Normalizer.}

\item{\code{entropy}}{numeric. The entropy value.}

\item{\code{min_max}}{data.frame. The minimum and maximum values of the continuous
variables, used for clamping.}
}}

\author{
Sergio Vignali
}
