% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxentTh.R
\name{maxentTh}
\alias{maxentTh}
\title{MaxEnt Thresholds}
\usage{
maxentTh(model)
}
\arguments{
\item{model}{\link{SDMmodel} object trained using the "Maxent" method.}
}
\description{
Returns the value of the thresholds generated by MaxEnt software.
}
\examples{
\dontrun{maxentTh(model)}
}
\author{
Sergio Vignali
}
