% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxentVarImp.R
\name{maxentVarImp}
\alias{maxentVarImp}
\title{Maxent Variable Importance}
\usage{
maxentVarImp(model)
}
\arguments{
\item{model}{SDMmodel object.}
}
\value{
A data frame with the variable importance.
}
\description{
Shows the percent contribution and permutation importance of the variable used to train the model.
}
\examples{
\dontrun{
maxentVarImp(model)}

}
\author{
Sergio Vignali
}
