% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Maxnet-class.R
\docType{class}
\name{Maxnet-class}
\alias{Maxnet-class}
\alias{Maxnet}
\alias{show,Maxnet-method}
\title{Maxnet}
\usage{
\S4method{show}{Maxnet}(object)
}
\arguments{
\item{object}{Maxnet object}
}
\description{
This Class represents a Maxnet model objects and hosts all the information
related to the model.
}
\section{Slots}{

\describe{
\item{\code{reg}}{numeric. The value of the regularization multiplier used to train
the model.}

\item{\code{fc}}{character. The feature class combination used to train the model.}

\item{\code{model}}{maxnet. The maxnet model object.}
}}

\author{
Sergio Vignali
}
