% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeSWD.R
\name{mergeSWD}
\alias{mergeSWD}
\title{Merge SWD Objects}
\usage{
mergeSWD(swd1, swd2)
}
\arguments{
\item{swd1}{\link{SWD} object.}

\item{swd2}{\link{SWD} object.}
}
\value{
The merged \link{SWD} object.
}
\description{
Merge two \link{SWD} object.
}
\details{
In case the two \link{SWD} objects have different columns, only the
common columns are used in the merged object.
}
\examples{
\dontrun{
mergeSWD(train, val)}

}
\author{
Sergio Vignali
}
