% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelOutput.R
\name{modelOutput}
\alias{modelOutput}
\title{Model Output}
\usage{
modelOutput(model, type = c("cloglog", "logistic", "raw"),
  response_curves = FALSE, jk = FALSE, threads = 1, folder = NULL)
}
\arguments{
\item{model}{Maxent object.}

\item{type}{character MaxEnt output type, possible values are "cloglog",
"logistic" and "raw", default is "cloglog".}

\item{response_curves}{logical, if TRUE it plots the response curves in the
html output, default is FALSE.}

\item{jk}{logical, if TRUE it runs the jackknife test, default FALSE.}

\item{threads}{numeric. Number of threads used by MaxEnt to compute the
Jackknife test, default is 1.}

\item{folder}{character. The name of the folder in which to save the MaxEnt
output, default is NULL meaning a temporary folder. The folder is created in
the working directory.}
}
\value{
The model object.
}
\description{
Given a model the function produces the output in the given folder.
}
\examples{
\dontrun{model <- modelOutput(model, type = "logistic", response_curves = T,
jk = T, thread = 4, folder = "my_folder"))}

}
\author{
Sergio Vignali
}
