% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NN_class.R
\docType{class}
\name{NN-class}
\alias{NN-class}
\alias{NN}
\title{NN}
\description{
This Class represents a Neural Network model objects and hosts all the information
related to the model.
}
\section{Slots}{

\describe{
\item{\code{model}}{keras.engine.sequential.Sequential. The Neural Network object.}

\item{\code{optimizer}}{character. Name of optimizer algorithm used to train the
model.}

\item{\code{loss}}{character. Name of the loss function optimized.}

\item{\code{means}}{numeric. Vector of means used to standardize continuous
variables.}

\item{\code{stds}}{numeric. Vector of standard deviations used to standardize
continuous variables.}
}}

\author{
Sergio Vignali
}
