% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NN_class.R
\docType{class}
\name{NN-class}
\alias{NN-class}
\alias{NN}
\title{NN}
\description{
This Class represents a Neural Network model objects and hosts all the information
related to the model.
}
\section{Slots}{

\describe{
\item{\code{model}}{keras.engine.sequential.Sequential. The Neural Network object.}
}}

\author{
Sergio Vignali
}
