% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimiseModel.R
\name{optimiseModel}
\alias{optimiseModel}
\title{Optimise Model}
\usage{
optimiseModel(model, bg4test, regs, fcs, bgs, test, pop, gen,
  metric = c("auc", "tss", "aicc"), env = NULL, parallel = FALSE,
  keep_best = 0.4, keep_random = 0.1, mutation_chance = 0.4,
  seed = NULL)
}
\arguments{
\item{model}{SDMmodel, SDMmodelCV object or a list containing models, see
details.}

\item{bg4test}{SWD object. Background locations used to get subsamples.}

\item{regs}{numeric vector with the regularization values to be tested.}

\item{fcs}{character vector with the feature combunation values to be tested.}

\item{bgs}{numeric vector with the number of background location to be
tested.}

\item{test}{SWD. Test dataset used to evaluate the model, not used with aicc
and SDMmodelCV objects, default is NULL.}

\item{pop}{numeric. Size of the population.}

\item{gen}{numeric. Number of generations.}

\item{metric}{character. The metric used to evaluate the models, possible
values are: "auc", "tss" and "aicc", default is "auc".}

\item{env}{\link{stack} or \link{brick} containing the environmental
variables, used only with "aicc", default is NULL.}

\item{parallel}{logical, if TRUE it uses parallel computation, deafult is
FALSE.}

\item{keep_best}{numeric. Percentage of the best models in the population to
be retained during each iteration, expressed as decimal number. Default
is 0.4.}

\item{keep_random}{numeric. Probability of retaining the excluded models
during each iteration, expressed as decimal number. Default is 0.1.}

\item{mutation_chance}{numeric. Probability of mutation of the child models,
expressed as decimal number. Default is 0.4.}

\item{seed}{numeric. The value used to set the seed to have consistent
results, default is NULL.}
}
\value{
list containing the models of the last population.
}
\description{
Optimise Model
}
\details{
You can continue to breed the same population of models passing the
output of the function (i.e. a list of models) as fist argument. See vignette
for a full explanation of the algorithm.
}
\examples{
\dontrun{output <- optimiseModel(my_model, bg, regs = c(0.5, 1,
1.5), fcs = c("lq", "lqp", "lqph"), bgs = c(5000, 10000, 15000),
test = my_val, pop = 20, gen = 10, seed = 25)}

}
\author{
Sergio Vignali
}
