% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permImp.R
\name{permImp}
\alias{permImp}
\title{Permutation Importance}
\usage{
permImp(model, permut = 10)
}
\arguments{
\item{model}{SDMsel object.}

\item{permut}{integer. Number of permutations, default is 10.}
}
\value{
data.frame with the ordered permutation importance.
}
\description{
The function randomly permutes one variable at time (using train and background
datasets) and computes the decrease in training AUC. The result is normalized
to percentages. Same implementation of MaxEnt java software but with the additional
possibility of running several permutations to obtain a better estimate of the
permutation importance. In case of more than one permutation (default is 10)
the average of the decrease in training AUC is computed.
}
\details{
Note that it could return values slightly different from MaxEnt java
software for a different random permutation.
}
\examples{
\dontrun{
permImp(model, permut = 50)}

}
\author{
Sergio Vignali
}
