% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDMtune-class.R
\docType{methods}
\name{plot,SDMtune,missing-method}
\alias{plot,SDMtune,missing-method}
\title{Plot SDMtune object}
\usage{
\S4method{plot}{SDMtune,missing}(x, title = "", interactive = FALSE)
}
\arguments{
\item{x}{\linkS4class{SDMtune} object.}

\item{title}{character. The title of the plot, by default is an empty string.}

\item{interactive}{logical, if TRUE plot an interactive chart, default is
\code{FALSE}.}
}
\value{
If \code{interactive = FALSE} the function returns a
\code{\link[ggplot2]{ggplot}} object otherwise it returns an SDMtuneChart
object that contains the path of the temporary folder where the necessary
files to create the chart are saved. This object can be used to saved the
chart in a file using the \link{saveChart} function. In both cases the
objects are returned as invisible.
}
\description{
Plot an \linkS4class{SDMtune} object. Use the interactive argument to create
an interactive chart.
}
\examples{
\donttest{
# Acquire environmental variables
files <- list.files(path = file.path(system.file(package = "dismo"), "ex"),
                    pattern = "grd", full.names = TRUE)
predictors <- raster::stack(files)

# Prepare presence locations
p_coords <- condor[, 1:2]

# Prepare background locations
bg_coords <- dismo::randomPoints(predictors, 5000)

# Create SWD object
presence <- prepareSWD(species = "Vultur gryphus", coords = p_coords,
                       env = predictors, categorical = "biome")
bg <- prepareSWD(species = "Vultur gryphus", coords = bg_coords,
                 env = predictors, categorical = "biome")

# Split presence locations in training (80\%) and testing (20\%) datasets
datasets <- trainValTest(presence, test = 0.2)
train <- datasets[[1]]
test <- datasets[[2]]

# Train a model
model <- train(method = "Maxnet", p = train, a = bg, fc = "l")

# Define the hyperparameters to test
h <- list(reg = 1:2, fc = c("lqp", "lqph"), a = c(1000, 2000))

# Run the gridSearch function using as metric the AUC
output <- gridSearch(model, hypers = h, metric = "auc", test = test,
                     bg4test = bg)

# Plot the output
plot(output, title = "My experiment")

# Plot the interactive chart
p <- plot(output, title = "My experiment", interactive = TRUE)
p
# Print the temporary folder that stores the files used to create the chart
str(p)
}
}
\seealso{
\link{saveChart}
}
\author{
Sergio Vignali
}
