% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDMtune_class.R
\docType{methods}
\name{plot,SDMtune,missing-method}
\alias{plot,SDMtune,missing-method}
\title{Plot}
\usage{
\S4method{plot}{SDMtune,missing}(x, interactive = FALSE)
}
\arguments{
\item{x}{SDMtune object.}
}
\description{
Plot the result of a tune function or of the optimiseModel function.
}
\examples{
\dontrun{plot(my_tuneBg)}

}
\author{
Sergio Vignali
}
