% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDMtune_class.R
\docType{methods}
\name{plot,SDMtune,missing-method}
\alias{plot,SDMtune,missing-method}
\title{Plot SDMtune object}
\usage{
\S4method{plot}{SDMtune,missing}(x, interactive = FALSE)
}
\arguments{
\item{x}{\link{SDMtune} object.}

\item{interactive}{logical, if TRUE plot an interactive chart.}
}
\description{
Plot an \link{SDMtune} object. Use the interactive argument to create an
interactive chart.
}
\examples{
\dontrun{plot(my_SDMtune_object)}

}
\author{
Sergio Vignali
}
