% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCor.R
\name{plotCor}
\alias{plotCor}
\title{Plot Correlation}
\usage{
plotCor(bg, method = "spearman", cor_th = NULL, text_size = 3)
}
\arguments{
\item{bg}{\linkS4class{SWD} object used to compute the correlation matrix.}

\item{method}{character. The method used to compute the correlation matrix.}

\item{cor_th}{numeric. If provided it prints only the coefficients that are
higher or lower than the given threshold.}

\item{text_size}{numeric, used to change the size of the text.}
}
\value{
A \link[ggplot2]{ggplot} object.
}
\description{
Plot a correlation matrix heat map with the value of the correlation
coefficients according with the given method. If cor_th is passed then it
prints only the coefficients that are higher or lower than the given
threshold.
}
\examples{
\donttest{# Acquire environmental variables
files <- list.files(path = file.path(system.file(package = "dismo"), "ex"),
                    pattern = "grd",
                    full.names = TRUE)

predictors <- terra::rast(files)

# Prepare background locations
bg_coords <- terra::spatSample(predictors,
                               size = 9000,
                               method = "random",
                               na.rm = TRUE,
                               xy = TRUE,
                               values = FALSE)

# Create SWD object
bg <- prepareSWD(species = "Virtual species",
                 a = bg_coords,
                 env = predictors,
                 categorical = "biome")

# Plot heat map
plotCor(bg,
        method = "spearman")

# Plot heat map showing only values higher than given threshold and change
# text size
plotCor(bg,
        method = "spearman",
        cor_th = 0.8,
        text_size = 4)}
}
\author{
Sergio Vignali
}
