% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotJk.R
\name{plotJk}
\alias{plotJk}
\title{Plot Jackknife Test}
\usage{
plotJk(jk, type = c("train", "test"), ref = NULL)
}
\arguments{
\item{jk}{data.frame with the output of the doJk function.}

\item{type}{character, "train" or "test" to plot the train or test AUC values.}

\item{ref}{numeric. The value of the chosen metric for the model trained using
all the variables. If provided it plots a vertical line showing the reference
value. Default is NULL.}
}
\value{
The ggplot object.
}
\description{
Plot the Jackknife Test for variable importance.
}
\examples{
\dontrun{
plotJk(jk_test, type = "train", ref = 0.955)}

}
\author{
Sergio Vignali
}
