% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPred.R
\name{plotPred}
\alias{plotPred}
\title{Plot Prediction}
\usage{
plotPred(map, lt = "", colorramp = NULL, hr = FALSE)
}
\arguments{
\item{map}{\link{raster} object with the prediction.}

\item{lt}{character. Legend title, default is an empty string.}

\item{colorramp}{vector. A custom color ramp given as a vector of colors
(see example), default is NULL and uses a color ramp similar to the original
MaxEnt output. See \href{color brewer}{http://colorbrewer2.org/} to make
nice color ramps.}

\item{hr}{logical, if TRUE produces an output with high resolution, default
is FALSE.}
}
\value{
A ggplot2 object.
}
\description{
Plot Prediction output.
}
\examples{
\dontrun{
plotPrediction(my_map, colorramp = c("#2c7bb6", "#abd9e9", "#ffffbf",
"#fdae61", "#d7191c"))}

}
\author{
Sergio Vignali
}
