% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotResponse.R
\name{plotResponse}
\alias{plotResponse}
\title{Plot Response Curve}
\usage{
plotResponse(
  model,
  var,
  type = NULL,
  only_presence = FALSE,
  marginal = FALSE,
  fun = mean,
  rug = FALSE,
  color = "red"
)
}
\arguments{
\item{model}{\linkS4class{SDMmodel} or \linkS4class{SDMmodelCV} object.}

\item{var}{character. Name of the variable to be plotted.}

\item{type}{character. The output type used for "Maxent" and "Maxnet"
methods, possible values are "cloglog" and "logistic".}

\item{only_presence}{logical. If \code{TRUE} it uses only the presence locations
when applying the function for the marginal response.}

\item{marginal}{logical. If \code{TRUE} it plots the marginal response curve.}

\item{fun}{function used to compute the level of the other variables for
marginal curves.}

\item{rug}{logical. If \code{TRUE} it adds the rug plot for the presence and
absence/background locations, available only for continuous variables.}

\item{color}{The color of the curve, default is "red".}
}
\value{
A \link[ggplot2]{ggplot} object.
}
\description{
Plot the Response Curve of the given environmental variable.
}
\details{
\itemize{
\item Note that \code{fun} is not a character argument, you must use \code{mean} and not
\code{"mean"}.
\item If you want to modify the plot, first you have to assign the output of the
function to a variable, and then you have two options:
\itemize{
\item Modify the \code{ggplot} object by editing the theme or adding additional
elements
\item Get the data with \code{ggplot2::ggplot_build()} and then build your own
plot (see examples)
}
}
}
\examples{
\donttest{# Acquire environmental variables
files <- list.files(path = file.path(system.file(package = "dismo"), "ex"),
                    pattern = "grd",
                    full.names = TRUE)

predictors <- terra::rast(files)

# Prepare presence and background locations
p_coords <- virtualSp$presence
bg_coords <- virtualSp$background

# Create SWD object
data <- prepareSWD(species = "Virtual species",
                   p = p_coords,
                   a = bg_coords,
                   env = predictors,
                   categorical = "biome")

# Train a model
model <- train(method = "Maxnet",
               data = data,
               fc = "lq")

# Plot cloglog response curve for a continuous environmental variable (bio1)
plotResponse(model,
             var = "bio1",
             type = "cloglog")

# Plot marginal cloglog response curve for a continuous environmental
# variable (bio1)
plotResponse(model,
             var = "bio1",
             type = "cloglog",
             marginal = TRUE)

# Plot logistic response curve for a continuous environmental variable
# (bio12) adding the rugs and giving a custom color
plotResponse(model,
             var = "bio12",
             type = "logistic",
             rug = TRUE,
             color = "blue")

# Plot response curve for a categorical environmental variable (biome) giving
# a custom color
plotResponse(model,
             var = "biome",
             type = "logistic",
             color = "green")

# Modify plot
# Change y axes limits
my_plot <- plotResponse(model,
                        var = "bio1",
                        type = "cloglog")
my_plot +
  ggplot2::scale_y_continuous(limits = c(0, 1))

# Get the data and create your own plot:
df <- ggplot2::ggplot_build(my_plot)$data[[1]]
plot(df$x, df$y,
     type = "l",
     lwd = 3,
     col = "blue",
     xlab = "bio1",
     ylab = "cloglog output")

# Train a model with cross validation
folds <- randomFolds(data,
                     k = 4,
                     only_presence = TRUE)

model <- train(method = "Maxnet",
               data = data,
               fc = "lq",
               folds = folds)

# Plot cloglog response curve for a continuous environmental variable (bio17)
plotResponse(model,
             var = "bio1",
             type = "cloglog")

# Plot logistic response curve for a categorical environmental variable
# (biome) giving a custom color
plotResponse(model,
             var = "biome",
             type = "logistic",
             color = "green")}
}
\author{
Sergio Vignali
}
