% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVarImp.R
\name{plotVarImp}
\alias{plotVarImp}
\title{Plot Varible Importance}
\usage{
plotVarImp(df, color = "grey")
}
\arguments{
\item{df}{data.frame. The result of varImp function.}

\item{color}{character. The color of the bar plot, default is grey.}
}
\value{
The ggplot object.
}
\description{
Plot the permutation variable importance as a bar plot.
}
\examples{
\dontrun{
plotVarImp(my_table, color = '#a1d99b')}

}
\author{
Sergio Vignali
}
