% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.Maxent}
\alias{predict.Maxent}
\title{Predict Maxent}
\usage{
predict.Maxent(model, data, clamp = TRUE, maxent_output = c("logistic",
  "cloglog", "raw"), filename = "", format = "GTiff", extent = NULL,
  parallel = FALSE)
}
\arguments{
\item{model}{A Maxent model object.}

\item{data}{A data frame, a \link\link{stack:raster} or \link{brick}.}

\item{clamp}{Flag for clumping during prediction, default is TRUE.}

\item{maxent_output}{The output type, possible values are "raw", "logistic" or "cloglog".}

\item{filename}{Output file name for the prediction map. If not provided the output is
saved in a file.}

\item{format}{The output format, see \link{writeRaster} for all the options, default is "GTiff".}

\item{extent}{\link{Extent} object, if provided it restricts the prediction to the given
extent, default is NULL.}

\item{parallel}{Flag to use parallel computation during prediction, default is FALSE.}
}
\value{
A vector of prediction or a Raster object if data is a raster stack.
}
\description{
Predict Maxent
}
\details{
You need package \link{snow} to use parallel computation.
}
\examples{
\dontrun{
predic(model, predictors, maxent_output = "cloglog", parallel = TRUE)}

}
\author{
Sergio Vignali
}
