% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_Maxent.R
\docType{methods}
\name{predict,Maxent-method}
\alias{predict,Maxent-method}
\title{Predict Maxent}
\usage{
\S4method{predict}{Maxent}(object, data, clamp = TRUE, filename = "",
  format = "GTiff", extent = NULL, parallel = FALSE, progress = "",
  type = NULL, ...)
}
\arguments{
\item{object}{Maxent object.}

\item{data}{data.frame, \link{SWD}, \link{stack} or \link{brick}.}

\item{clamp}{logical for clumping during prediction, default is TRUE.}

\item{filename}{character. Output file name for the prediction map, if provided the output is
saved in a file.}

\item{format}{character. The output format, see \link{writeRaster} for all the options, default is "GTiff".}

\item{extent}{\link{Extent} object, if provided it restricts the prediction to the given
extent, default is NULL.}

\item{parallel}{logical to use parallel computation during prediction, default is FALSE.}

\item{progress}{character to display a progress bar: "text", "window" or "" (default) for no progress bar.}

\item{type}{character MaxEnt output type, if not provided it uses the model type.
Possible values are "cloglog", "logistic" and "raw", default is NULL.}

\item{...}{Additional parameter to pass to the \link{writeRaster} function.}
}
\value{
A vector of prediction or a Raster object if data is a raster stack/brick.
}
\description{
Compute prediction for a new dataset from a trained Maxent model.
}
\details{
You need package \pkg{snow} to use parallel computation and \pkg{rgdal}
to save the prediction in a raster file. Parallel computation increases the speed
only for big datasets due to the time necessary to create the cluster.
The function performs the prediction in \strong{R} without calling the \strong{MaxEnt}
java software. This results is a faster computation for large datasets.
}
\examples{
\dontrun{
predict(model, predictors, parallel = TRUE)}

}
\references{
\insertRef{Wilson2009}{SDMsel}
}
\author{
Sergio Vignali
}
