% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareSWD.R
\name{prepareSWD}
\alias{prepareSWD}
\title{Prepare a MaxentSWD data set for MaxEnt models}
\usage{
prepareSWD(species, coords, env, categoricals = NULL)
}
\arguments{
\item{species}{The name of the species, background in the case of background locations.}

\item{coords}{The coordinates of the presence or background locations.}

\item{env}{The environmental variables in a RasterStack format.}

\item{categoricals}{Vector indicating which of the environmental variable is categorical, default is NULL.}
}
\value{
A MaxentSWD object
}
\description{
Given the coordinates, the species' name and the environmental variables,
the function prepares a data frame in the SWD format (sample with data).
}
\examples{
\dontrun{
presence_swd <- prepareSWD(species = "Bradipus variegatus", coords = presence, env = env)}

}
\author{
Sergio Vignali
}
