% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareSWD.R
\name{prepareSWD}
\alias{prepareSWD}
\title{Prepare a SWD data set for MaxEnt models}
\usage{
prepareSWD(species, coords, env, categoricals = NULL)
}
\arguments{
\item{species}{character. The name of the species.}

\item{coords}{data.frame. The coordinates of the presence or background
locations.}

\item{env}{\link{stack} or \link{brick} containing the environmental
variables used to train the model.}

\item{categoricals}{vector indicating which of the environmental variable are
categoricals, default is NULL.}
}
\value{
A SWD object
}
\description{
Given the coordinates, the species' name and the environmental variables,
the function prepares a data frame in the SWD format (sample with data).
}
\examples{
\dontrun{
presence_swd <- prepareSWD(species = "Bradipus variegatus",
coords = presence, env = env)}

}
\author{
Sergio Vignali
}
