% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareSWD.R
\name{prepareSWD}
\alias{prepareSWD}
\title{Prepare an SWD object}
\usage{
prepareSWD(species, p = NULL, a = NULL, coords = NULL, env,
  categorical = NULL)
}
\arguments{
\item{species}{character. The name of the species.}

\item{p}{data.frame. The coordinates of the presence locations.}

\item{a}{data.frame. The coordinates of the absence/background locations.}

\item{coords}{Deprecated.}

\item{env}{\code{\link[raster]{stack}} containing the environmental variables
used to extract the values at coordinate locations.}

\item{categorical}{vector indicating which of the environmental variable are
categorical, default is \code{NULL}.}
}
\value{
An \linkS4class{SWD} object.
}
\description{
Given the coordinates, the species' name and the environmental variables,
the function creates an \linkS4class{SWD} object (sample with data).
}
\examples{
# Acquire environmental variables
files <- list.files(path = file.path(system.file(package = "dismo"), "ex"),
                    pattern = "grd", full.names = TRUE)
predictors <- raster::stack(files)

# Prepare presence locations
p_coords <- condor[, 1:2]

# Prepare background locations
set.seed(25)
bg_coords <- dismo::randomPoints(predictors, 2000)

train <- prepareSWD(species = "Vultur gryphus", p = p_coords, a = bg_coords,
                    env = predictors, categorical = "biome")
presence
}
\author{
Sergio Vignali
}
