% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\docType{methods}
\name{response,Maxent-method}
\alias{response,Maxent-method}
\title{Response Curve}
\usage{
\S4method{response}{Maxent}(object, variable, marginal = FALSE,
  clamp = TRUE, rug = FALSE, color = "red")
}
\arguments{
\item{object}{Maxent object.}

\item{variable}{character. Name of the variable to be plotted.}

\item{marginal}{logical, if TRUE it plots the marginal response courve, default is FALSE.}

\item{clamp}{logical for clumping during prediction, default is TRUE.}

\item{rug}{logical, if TRUE it adds the rug plot for presence and background locations locations,
available only for continuous variables,
default is FALSE.}

\item{color}{The color of the curve, default is "red".}
}
\value{
The plot objects
}
\description{
Plot the Response Curve of the given environmental variable.
}
\examples{
\dontrun{
response(model, variable = "bio12", marginal = TRUE, rug = TRUE)}

}
\author{
Sergio Vignali
}
