% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RF-class.R
\docType{class}
\name{RF-class}
\alias{RF-class}
\alias{RF}
\title{Random Forest}
\description{
This Class represents a Random Forest model objects and hosts all the
information related to the model.
}
\details{
See \code{\link{randomForest}} for the meaning of the slots.
}
\section{Slots}{

\describe{
\item{\code{mtry}}{integer. Number of variable randomly sampled.}

\item{\code{ntree}}{integer. Number of grown trees.}

\item{\code{nodesize}}{integer. Minimum size of terminal nodes.}

\item{\code{model}}{\code{\link{randomForest}}. The randomForest model object.}
}}

\author{
Sergio Vignali
}
