% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDMmodel_class.R
\docType{class}
\name{SDMmodel-class}
\alias{SDMmodel-class}
\alias{SDMmodel}
\title{SDMmodel}
\description{
This Class represents a SDMmodel model object and hosts all the information
related to the model.
}
\section{Slots}{

\describe{
\item{\code{p}}{\link{SWD} object. The presence locations used to train the model.}

\item{\code{a}}{\link{SWD} object. The absence or background locations used to train
the model.}

\item{\code{model}}{\link{Maxent} or \link{Maxnet} object.}
}}

\author{
Sergio Vignali
}
