% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDMmodel-class.R
\docType{class}
\name{SDMmodel-class}
\alias{SDMmodel-class}
\alias{SDMmodel}
\title{SDMmodel}
\description{
This Class represents an SDMmodel model object and hosts all the information
related to the model.
}
\section{Slots}{

\describe{
\item{\code{data}}{\code{\linkS4class{SWD}} object. The data used to train the model.}

\item{\code{model}}{\link{Maxent} or \link{Maxnet} object.}
}}

\author{
Sergio Vignali
}
