% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDMmodel-class.R
\docType{class}
\name{SDMmodel-class}
\alias{SDMmodel-class}
\alias{SDMmodel}
\alias{show,SDMmodel-method}
\title{SDMmodel}
\usage{
\S4method{show}{SDMmodel}(object)
}
\arguments{
\item{object}{SDMmodel object}
}
\description{
This Class represents an SDMmodel object and hosts all the information
related to the model.
}
\section{Slots}{

\describe{
\item{\code{data}}{\linkS4class{SWD} object. The data used to train the model.}

\item{\code{model}}{An object of class \linkS4class{ANN}, \linkS4class{BRT},
\linkS4class{RF}, \linkS4class{Maxent} or \linkS4class{Maxnet}.}
}}

\author{
Sergio Vignali
}
