% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDMmodel2MaxEnt.R
\name{SDMmodel2MaxEnt}
\alias{SDMmodel2MaxEnt}
\title{SDMmodel2MaxEnt}
\usage{
SDMmodel2MaxEnt(model)
}
\arguments{
\item{model}{SDMmodel object to be converted.}
}
\value{
The converted MaxEnt object.
}
\description{
Converts a SDMmodel object containing a Maxent model to a dismo MaxEnt
object.
}
\examples{
\dontrun{SDMmodel2MaxEnt(model)}

}
\author{
Sergio Vignali
}
