% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDMmodelCV.R
\docType{class}
\name{SDMmodelCV-class}
\alias{SDMmodelCV-class}
\alias{SDMmodelCV}
\title{SDMmodelCV}
\description{
This Class represents a SDMmodel model object with replicates and hosts all
the models trained during the cross validation.
}
\section{Slots}{

\describe{
\item{\code{models}}{list. A list containing all the models trained during the cross
validation.}

\item{\code{presence}}{SWD object. Full dataset used by the k-fold cross validation.}

\item{\code{folds}}{numeric. Vector with the index of the k-fold.}
}}

\author{
Sergio Vignali
}
