% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDMmodelCV-class.R
\docType{class}
\name{SDMmodelCV-class}
\alias{SDMmodelCV-class}
\alias{SDMmodelCV}
\title{SDMmodelCV}
\description{
This Class represents a SDMmodel model object with replicates and hosts all
the models trained during the cross validation.
}
\section{Slots}{

\describe{
\item{\code{models}}{list. A list containing all the models trained during the cross
validation.}

\item{\code{p}}{\link{SWD} object. Full dataset used by the k-fold cross validation.}

\item{\code{a}}{\link{SWD} object. The absence/background locations used to train the
model.}

\item{\code{folds}}{list with two matices, the first for the training and the second
for the testing dataset. Each column of one metrix represents a fold with
\code{TRUE} for train locations and \code{FALSE} for test locations.}
}}

\author{
Sergio Vignali
}
