% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDMmodelCV-class.R
\docType{class}
\name{SDMmodelCV-class}
\alias{SDMmodelCV-class}
\alias{SDMmodelCV}
\alias{show,SDMmodelCV-method}
\title{SDMmodelCV}
\usage{
\S4method{show}{SDMmodelCV}(object)
}
\arguments{
\item{object}{SDMmodelCV object}
}
\description{
This Class represents an SDMmodel model object with replicates and hosts all
the models trained during the cross validation.
}
\section{Slots}{

\describe{
\item{\code{models}}{list. A list containing all the models trained during the cross
validation.}

\item{\code{data}}{\linkS4class{SWD} object. Full dataset used to make the
partitions.}

\item{\code{folds}}{list with two matrices, the first for the training and the second
for the testing dataset. Each column of one matrix represents a fold with
\code{TRUE} for the locations included in and \code{FALSE} excluded from the partition.}
}}

\author{
Sergio Vignali
}
