% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDMsel.R
\docType{package}
\name{SDMsel}
\alias{SDMsel}
\alias{SDMsel-package}
\title{SDMsel: A package for tuning Species Distribution Models.}
\description{
Performs data driven variable selection and model tuning for SDMs
(Species Distribution Models). It provides also some utilities to display
results (at the moment it implements only MaxEnt models).
}
\details{
The foo functions ...
}
