% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholds.R
\name{thresholds}
\alias{thresholds}
\title{Thresholds}
\usage{
thresholds(model, type, test = NULL)
}
\arguments{
\item{model}{\link{SDMmodel} object.}

\item{type}{character. The output type, possible values are "cloglog" and
"logistic", default is "cloglog".}

\item{test}{\link{SWD} test locations, if not provided it returns the
training and test thresholds, default is NULL.}
}
\value{
data.frame with the thresholds.
}
\description{
Compute three threshold values: minimum training presence, equal training
sensitivity and specificity and maximum training sensitivity plus
specificity together with fractional predicted area and the omission rate. If
a test dataset is provided it returns also the equal test sensitivity and
specificity and maximum test sensitivity plus specificity thresholds and the
p-values of the one-tailed binomial exact test.
}
\details{
The equal training sensitivity and specificity minimizes the
difference between sensitivity and specificity. The one-tailed binomial test
checks that test points are predicted no better than by a random prediction
with the same fractional predicted area.
}
\examples{
\dontrun{thresholds(model, type = "logistic")}

}
\author{
Sergio Vignali
}
