% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{train}
\alias{train}
\title{Train}
\usage{
train(method = c("Maxent", "Maxnet"), p, a, rep = 1, verbose = TRUE,
  folds = NULL, seed = NULL, ...)
}
\arguments{
\item{method}{character. Possible values are "Maxent" or "Maxnet".}

\item{p}{\link{SWD} object with the presence locations.}

\item{a}{\link{SWD} object with the absence or background locations.}

\item{rep}{numeric. Number of replicates, used for cross validation.
Default is 1, meaning no cross validation is performed.}

\item{verbose}{logical if TRUE shows a progress bar if replicates is greater
then 1, default is TRUE.}

\item{folds}{numeric. Vector containing the indexes for the k-fold partition
of the training data, if not provided the function randomly creates the
folds, default is NULL.}

\item{seed}{integer. The value used to set the seed for the fold partition,
used if \strong{folds} is not provided, default is NULL.}

\item{...}{Arguments passed to the relative functions, see \link{trainMaxent}
or \link{trainMaxnet} for details related to the different methods.}
}
\value{
A \link{SDMmodel} or \link{SDMmodelCV} object.
}
\description{
Train a model using the given method.
}
\examples{
\dontrun{
model <- train("Maxnet", p, a, reg = 2, fc = "lqp")}

}
\author{
Sergio Vignali
}
