% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{train}
\alias{train}
\title{Train}
\usage{
train(method, data = NULL, p = NULL, a = NULL, rep = 1,
  verbose = TRUE, folds = NULL, seed = NULL, ...)
}
\arguments{
\item{method}{character. Method used to train the model, possible values are
"Maxent" or "Maxnet".}

\item{data}{\code{\linkS4class{SWD}} object with presence and
absence/background locations.}

\item{p}{Deprecated.}

\item{a}{Deprecated.}

\item{rep}{numeric. Number of replicates, used for cross validation.
Default is 1, meaning no cross validation is performed.}

\item{verbose}{logical, if \code{TRUE} shows a progress bar during cross
validation (i.e. when \code{rep > 1}), default is \code{TRUE}.}

\item{folds}{matrix. Each column representing a fold with \code{TRUE} for
train locations and \code{FALSE} for test locations.}

\item{seed}{Deprecated.}

\item{...}{Arguments passed to the relative method, see details.}
}
\value{
An \code{\linkS4class{SDMmodel}} or \code{\linkS4class{SDMmodelCV}}
object.
}
\description{
Train a model using the given method.
}
\details{
\itemize{
\item For the Maxent method, possible arguments are:
\itemize{
\item reg: numeric. The value of the regularization multiplier, default is 1.
\item fc: vector. The value of the feature classes, possible values are
combinations of "l", "q", "p", "h" and "t", default is "lqph".
\item iter: numeric. Number of iterations used by the MaxEnt algorithm,
default is 500.
\item extra_args: vector. Extra arguments used to run MaxEnt, default is
"removeduplicates=false" and "addsamplestobackground=false". In case
this is not your expected behavior you can assign extra_args = "" or
you can change or add any other additional arguments extending the
default settings (e.g. \code{extra_args = c("removeduplicates=true,
      addsamplestobackground=true)"**})
}
\item For the Maxnet method, possible arguments are:
\itemize{
\item reg: numeric. The value of the regularization intensity, default is 1.
\item fc: vector. The value of the feature classes, possible values are
combinations of "l", "q", "p", "h" and "t", default is "lqph". For more
details see \code{\link[maxnet]{maxnet}}.
}
\item The length of each column in the \code{fold} argument must be the same of
the locations used to train the model.
}
}
\examples{
\donttest{
# Acquire environmental variables
files <- list.files(path = file.path(system.file(package = "dismo"), "ex"),
                    pattern = "grd", full.names = TRUE)
predictors <- raster::stack(files)

# Prepare presence locations
p_coords <- condor[, 1:2]

# Prepare background locations
bg_coords <- dismo::randomPoints(predictors, 5000)

# Create SWD object
train <- prepareSWD(species = "Vultur gryphus", p = p_coords, a = bg_coords,
                    env = predictors, categorical = "biome")

# Train a Maxent model
model <- train(method = "Maxent", data = train, fc = "l", reg = 1.5,
               iter = 700)

# Train a Maxnet model
model <- train(method = "Maxnet", data = train, fc = "lq", reg = 1.5)

# Train a Maxnet model with cross validation
model <- train(method = "Maxnet", data = train, fc = "l", reg = 0.8, rep = 4)
}
}
\seealso{
\code{\link{randomFolds}}
}
\author{
Sergio Vignali
}
