% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{train}
\alias{train}
\title{Train}
\usage{
train(method = c("Maxent", "Maxnet"), presence, bg, rm, fc, iter = 500,
  extra_args = c("noaddsamplestobackground", "removeduplicates=false"))
}
\arguments{
\item{method}{character. Possible values are "Maxent" or "Maxnet}

\item{presence}{SWD object with the presence locations.}

\item{bg}{SWD object with the background locations.}

\item{rm}{numeric. The value of the regularization multiplier.}

\item{fc}{vector. The value of the feature combination, possible values are
combinations of "l", "q", "p", "h" and "t".}

\item{iter}{numeric. Number of iterations used by the Maxent alghoritm,
default is 500, used only for "Maxent" method.}

\item{extra_args}{vector. Extra arguments used to run MaxEnt, defalt is
c("noaddsamplestobackground", "removeduplicates=false"), used only for
"Maxent" method.#'}
}
\value{
A SWDmodel object
}
\description{
Train a model using the given method.
}
\details{
See \link{trainMaxent} and \link{trainMaxnet} for details related to
the different methods. For \strong{Maxent} models the function uses by default
\strong{extra_args = c("noaddsamplestobackground", "removeduplicates=false")}.
In case this is not your expected beaviour you can remove both passing
\strong{extra_args = ""} or you can add any other additional arguments extending
the previous vector.
}
\examples{
\dontrun{
model <- train("Maxnet", presence, bg, rm = 2, fc = "lqp")}

}
\author{
Sergio Vignali
}
