% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainMaxent.R
\name{trainMaxent}
\alias{trainMaxent}
\title{Train Maxent model}
\usage{
trainMaxent(presence, bg, rm, fc, test = NULL, maxent_output = "cloglog",
  iterations = 500, extra_args = NULL, folder = NULL)
}
\arguments{
\item{presence}{SWD object with the presence locations.}

\item{bg}{SWD object with the background locations.}

\item{rm}{numeric. The value of the regularization multiplier.}

\item{fc}{vector. The value of the feature combination, possible values are combinations of
"l", "q", "p", "h" and "t".}

\item{test}{SWD object with the test locations, default is NULL.}

\item{maxent_output}{character. The MaxEnt output format, possible values are "logistic", "cloglog"
and "raw", default value "cloglog".}

\item{iterations}{numeric. Number of iterations used by the Maxent alghoritm, default is 500.}

\item{extra_args}{vector. Extra arguments used to run MaxEnt, e.g. "removeduplicates=false", default
is NULL.}

\item{folder}{character. The folder name where to save the MaxEnt output, default is NULL meaning
that is not saved. The folder is created in the working directory.}
}
\value{
The output of MaxEnt as Maxent object.
}
\description{
Trtain a MaxEnt model using the dismo package.
}
\examples{
\dontrun{model <- runMaxent(presence, bg, rm, response_curves = TRUE))}

}
\author{
Sergio Vignali
}
