% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainMaxent.R
\name{trainMaxent}
\alias{trainMaxent}
\title{Train Maxent model}
\usage{
trainMaxent(presence, bg, reg = 1, fc = "lqph", iter = 500,
  extra_args = "removeduplicates=false")
}
\arguments{
\item{presence}{SWD object with the presence locations.}

\item{bg}{SWD object with the background locations.}

\item{reg}{numeric. The value of the regularization multiplier, default is 1.}

\item{fc}{vector. The value of the feature combination, possible values are
combinations of "l", "q", "p", "h" and "t", default is "lqph".}

\item{iter}{numeric. Number of iterations used by the Maxent alghoritm,
default is 500.}

\item{extra_args}{vector. Extra arguments used to run MaxEnt, default is
"removeduplicates=false".}
}
\value{
A SDMmodel object.
}
\description{
Train a MaxEnt model using the dismo package. The function runs Maxent model
with a minimum setting of optional arguments (e.g. it doesn't make the
response curves) to speed up the computation time. Use \link{modelReport}
function to set more arguments (i.e. response curves) and save the output
permanently in a folder.
}
\details{
By default the function uses
\strong{extra_args = "removeduplicates=false"}. In case this is not your expected
beaviour you can assign extra_args = "" or you can add any other additional
arguments extending the previous vector.
}
\examples{
\dontrun{model <- trainMaxent(presence, bg)}

}
\author{
Sergio Vignali
}
