% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainMaxent.R
\name{trainMaxent}
\alias{trainMaxent}
\title{Train Maxent model}
\usage{
trainMaxent(p, a, reg = 1, fc = "lqph", iter = 500,
  extra_args = c("removeduplicates=false",
  "addsamplestobackground=false"))
}
\arguments{
\item{p}{\link{SWD} object with the presence locations.}

\item{a}{\link{SWD} object with the background locations.}

\item{reg}{numeric. The value of the regularization multiplier, default is 1.}

\item{fc}{vector. The value of the feature combination, possible values are
combinations of "l", "q", "p", "h" and "t", default is "lqph".}

\item{iter}{numeric. Number of iterations used by the MaxEnt algorithm,
default is 500.}

\item{extra_args}{vector. Extra arguments used to run MaxEnt, default is
"removeduplicates=false" and "addsamplestobackground=false".}
}
\value{
A \link{SDMmodel} object.
}
\description{
Train a \link{Maxent} model using the dismo package. The function runs Maxent
model with a minimum setting of optional arguments (e.g. it doesn't make the
response curves) to speed up the computation time. Use \link{modelReport}
function to set more arguments (i.e. response curves) and save the output
permanently in a folder.
}
\details{
By default the function uses
\strong{extra_args = c("removeduplicates=false, addsamplestobackground=false)"}.
In case this is not your expected behavior you can assign extra_args = "" or
you can add any other additional arguments extending the previous vector.
}
\examples{
\dontrun{model <- trainMaxent(p, a)}

}
\author{
Sergio Vignali
}
