% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainMaxnet.R
\name{trainMaxnet}
\alias{trainMaxnet}
\title{Train Maxnet model}
\usage{
trainMaxnet(presence, bg, rm, fc)
}
\arguments{
\item{presence}{SWD object with the presence locations.}

\item{bg}{SWD object with the background locations.}

\item{rm}{numeric. The value of the regularization multiplier.}

\item{fc}{vector. The value of the feature combination, possible values are
combinations of "l", "q", "p", "h" and "t".}
}
\value{
A SDMmodel object.
}
\description{
Train a Maxnet model using the maxnet package.
}
\examples{
\dontrun{model <- trainMaxnet(presence, bg, rm)}

}
\author{
Sergio Vignali
}
