% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainMaxnet.R
\name{trainMaxnet}
\alias{trainMaxnet}
\title{Train Maxnet model}
\usage{
trainMaxnet(p, a, reg = 1, fc = "lqph")
}
\arguments{
\item{p}{\link{SWD} object with the presence locations.}

\item{a}{\link{SWD} object with the background locations.}

\item{reg}{numeric. The value of the regularization intensity, default is 1.}

\item{fc}{vector. The value of the feature classes, possible values are
combinations of "l", "q", "p", "h" and "t", default is "lqph".}
}
\value{
A \link{SDMmodel} object.
}
\description{
Train a \link{Maxnet} model using the \link{maxnet} package.
}
\examples{
\dontrun{model <- trainMaxnet(p, a)}

}
\author{
Sergio Vignali
}
