% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainNN.R
\name{trainNN}
\alias{trainNN}
\title{Train Neural Network}
\usage{
trainNN(presence, bg, conf = NULL, model = NULL, reg = 0,
  optimizer = "rmsprop", loss = "mse", epoch = 500, batch_size = 128,
  verbose = 1, callbacks = list(), standardize = TRUE)
}
\arguments{
\item{standardize}{}
}
\description{
Train Neural Network
}
