% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainValTest.R
\name{trainValTest}
\alias{trainValTest}
\title{Train, Validation and Test datasets}
\usage{
trainValTest(x, test, val = 0, seed = NULL)
}
\arguments{
\item{x}{\link{SWD} object containing the data that have to be split in
training, validation and testing datasets.}

\item{test}{numeric. The percentage of data withhold for testing.}

\item{val}{numeric. The percentage of data withhold for validation, default
is 0.}

\item{seed}{numeric. The value used to set the seed in order to have
consistent results, default is NULL.}
}
\value{
A list with the training, validation and testing data sets or
training and testing data sets accordingly.
}
\description{
Split a dataset randomly in training and testing datasets or training,
validation and testing datasets.
}
\examples{
\dontrun{
sets <- trainValTest(p, val = 0.2, test = 0.2, set_seed = 25)}

}
\author{
Sergio Vignali
}
