% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tss.R
\name{tss}
\alias{tss}
\title{True Skill Statistics}
\usage{
tss(model, test = NULL)
}
\arguments{
\item{model}{Maxent object.}

\item{test}{SWD object, if provided it computes the TSS for the test dataset,
if not for the train dataset, default is NULL.}
}
\value{
The value of the TSS of the given model.
}
\description{
Compute the max TSS of a given Maxent model.
}
\examples{
\dontrun{
tss(model, test = test_dataset)}

}
\references{
Allouche O., Tsoar A., Kadmon R., (2006). Assessing the accuracy of species distribution models:
prevalence, kappa and the true skill statistic (TSS). Journal of Applied Ecology, 43(6), 1223–1232.
}
\author{
Sergio Vignali
}
