% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tss.R
\name{tss}
\alias{tss}
\title{True Skill Statistics}
\usage{
tss(model, test = NULL)
}
\arguments{
\item{model}{SDMmodel or SDMmodelCV object.}

\item{test}{SWD test locations for SDMmodel objects or TRUE for SDModelCV
objects, if not provided it computes the train AUC, default is NULL.}
}
\value{
The value of the TSS of the given model.
}
\description{
Compute the max TSS of a given model.
}
\details{
If the model is a SDMmodelCV object, the function computes the mean
of the training or testing TSS values of the different replicates.
}
\examples{
\dontrun{
tss(model, test = test_dataset)}

}
\references{
Allouche O., Tsoar A., Kadmon R., (2006). Assessing the accuracy of species distribution models:
prevalence, kappa and the true skill statistic (TSS). Journal of Applied Ecology, 43(6), 1223–1232.
}
\author{
Sergio Vignali
}
