% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuneBg.R
\name{tuneBg}
\alias{tuneBg}
\title{Tune number of Background locations}
\usage{
tuneBg(model, bg4test, bgs, metric = c("auc", "tss", "aicc"), env = NULL,
  parallel = FALSE, extra_args = NULL, seed = NULL)
}
\arguments{
\item{model}{Maxent object.}

\item{bg4test}{SWD. The dataset with the maximum number of background locations to be tested given as MaxentSWD object.}

\item{bgs}{vector. A sequence of number of background locations to test. The maximum number must be lower thatn the maximum number of
background locations in bg4test. Default is the maximum number of background locations in bg4test.}

\item{metric}{character. The metric used to evaluate the models, possible values are:
"auc", "tss" and "aicc", default is "auc"}

\item{env}{\link{stack} or \link{brick} containing the environmental variables,
used only with "aicc", default is NULL.}

\item{parallel}{logical, if TRUE it uses parallel computation, deafult is FALSE.}

\item{extra_args}{vector Extra arguments used to run MaxEnt..}

\item{seed}{integer. The value used to set the seed in order to have consistent results, default is NULL.}
}
\value{
A \link{SDMsel} object.
}
\description{
Test different sizes of background locations.
}
\examples{
\dontrun{
bg_test <- tuneBg(model, bg4test,bgs = seq(5000, 30000, 5000), metric = "auc", seed = 25)}

}
\author{
Sergio Vignali
}
