% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuneFC.R
\name{tuneFC}
\alias{tuneFC}
\title{Tune Feature Combination}
\usage{
tuneFC(model, fcs, metric = c("auc", "tss", "aicc"), env = NULL,
  parallel = FALSE)
}
\arguments{
\item{model}{Maxent object.}

\item{fcs}{vector. A list of Feature Classes combination to be tested.}

\item{metric}{character. The metric used to evaluate the models, possible values are:
"auc", "tss" and "aicc", default is "auc".}

\item{env}{\link{stack} or \link{brick} containing the environmental variables,
used only with "aicc", default is NULL.}

\item{parallel}{logical, if TRUE it uses parallel computation, deafult is FALSE.}
}
\value{
A \link{SDMsel} object.
}
\description{
The function iterates among different Feature Classes combinations and for each of them
performs a run a MaxEnt model using the given regularization multiplier.
}
\details{
You need package \pkg{snow} to use parallel computation and \pkg{rgdal}
to save the prediction in a raster file. Parallel computation increases the speed
only for big datasets due to the time necessary to create the cluster.
}
\examples{
\dontrun{
fc_test <- tuneFC(model, fcs = c("l", "lq", "lqp"), metric = "aicc",
env = predictors, parallel = T)}

}
\author{
Sergio Vignali
}
