% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varSel.R
\name{varSel}
\alias{varSel}
\title{Variable Selection}
\usage{
varSel(model, metric, bg4cor, test = NULL, env = NULL,
  parallel = FALSE, method = "spearman", cor_th = 0.7, permut = 10,
  use_pc = FALSE)
}
\arguments{
\item{model}{\link{SDMmodel} or \link{SDMmodelCV} object.}

\item{metric}{character. The metric used to evaluate the models, possible
values are: "auc", "tss" and "aicc".}

\item{bg4cor}{\link{SWD} object. Background locations used to test the
correlation between environmental variables.}

\item{test}{\link{SWD}. Test dataset used to evaluate the model, not used
with aicc and \link{SDMmodelCV} objects, default is NULL.}

\item{env}{\link{stack} containing the environmental variables, used only
with "aicc", default is NULL.}

\item{parallel}{logical, if TRUE it uses parallel computation, default is
FALSE. Used only with AICc.}

\item{method}{character. The method used to compute the correlation matrix,
default "spearman".}

\item{cor_th}{numeric. The correlation threshold used to select highly
correlated variables, default is 0.7.}

\item{permut}{integer. Number of permutations, default is 10.}

\item{use_pc}{logical, use percent contribution. If TRUE and the model is
trained using the \link{Maxent} method, the algorithm uses the percent
contribution computed by Maxent software to score the variable importance,
default is FALSE.}
}
\value{
The \link{SDMmodel} or \link{SDMmodelCV} object trained using the
selected variables.
}
\description{
The function performs a data-driven variable selection. Starting from the
provided model it iterates through all the variables starting from the one
with the highest contribution (permutation importance or maxent percent
contribution). If the variable is correlated with other variables (according
to the given method and threshold) it performs a Jackknife test and among the
correlated variables it removes the one that results in the best performing
model when removed (according to the given metric). The process is repeated
until the remaining variables are not highly correlated anymore.
}
\details{
Parallel computation increases the speed only for large datasets due
to the time necessary to create the cluster.\cr
To find highly correlated variables the following formula is used:
\deqn{| coeff | \le cor_th}
}
\examples{
\dontrun{
varSel(model, bg, metric = "auc")}

}
\author{
Sergio Vignali
}
